Tighten custom domain integration
