+++ apps/os/src/domains/repos/project-repo-template.generated.ts
@@ -8,7 +8,7 @@ export const PROJECT_REPO_INITIAL_FILES: Array<{ content: string; path: string }
   { path: "apps/counter/worker.ts", content: "import { DurableObject } from \"cloudflare:workers\";\n\n// A stateful app: a Durable Object class hosted as a repo-backed stateful\n// dynamic worker. State survives across requests under its durableWorkerKey.\nexport class CounterApp extends DurableObject {\n  async fetch(req: Request): Promise<Response> {\n    // The path lane advertises its stripped URL prefix; host lanes have none.\n    const prefix = req.headers.get(\"x-iterate-url-prefix\") ?? \"\";\n    const url = new URL(req.url);\n\n    if (req.method === \"POST\" && url.pathname === \"/increment\") {\n      await this.increment();\n      // Redirect back so the browser URL never sticks at /increment.\n      return new Response(null, { status: 303, headers: { location: `${prefix}/` } });\n    }\n\n    const count = await this.current();\n    return new Response(\n      `<!doctype html>\n        <html>\n          <body>\n            <main>\n              <p>count: ${count}</p>\n              <form method=\"post\" action=\"${prefix}/increment\">\n                <button type=\"submit\">increment</button>\n              </form>\n            </main>\n          </body>\n        </html>`,\n      { headers: { \"content-type\": \"text/html; charset=utf-8\" } },\n    );\n  }\n\n  async increment(): Promise<number> {\n    const n = (this.ctx.storage.kv.get<number>(\"n\") ?? 0) + 1;\n    this.ctx.storage.kv.put(\"n\", n);\n    return n;\n  }\n\n  async current(): Promise<number> {\n    return this.ctx.storage.kv.get<number>(\"n\") ?? 0;\n  }\n}\n" },
   { path: "apps/hello/worker.ts", content: "import { WorkerEntrypoint } from \"cloudflare:workers\";\nimport type { ProjectRpcTarget } from \"../../sdk.ts\";\n\n// A stateless app: a plain WorkerEntrypoint the root project worker routes\n// to when ingress selects the \"hello\" app. It still gets the full project\n// itx through env.ITX. Built from a masked snapshot of this repo (see the\n// APPS map in the root worker.ts), so it can import shared modules like\n// sdk.ts.\nexport default class HelloApp extends WorkerEntrypoint<{\n  ITX: { get(): Promise<ProjectRpcTarget> };\n}> {\n  async fetch(req: Request): Promise<Response> {\n    const project = await this.env.ITX.get();\n    const description = await project.__describe();\n    return Response.json({\n      app: \"hello\",\n      path: new URL(req.url).pathname,\n      projectId: description.projectId,\n    });\n  }\n}\n" },
   { path: "package.json", content: "{\n  \"name\": \"iterate-project-worker\",\n  \"private\": true,\n  \"version\": \"0.0.0\",\n  \"type\": \"module\",\n  \"description\": \"Iterate project worker. Dependencies listed here are installed by the worker build pipeline when the worker is bundled.\",\n  \"dependencies\": {\n    \"@slack/web-api\": \"^7.14.1\"\n  },\n  \"devDependencies\": {\n    \"@cloudflare/workers-types\": \"^4.20250620.0\",\n    \"typescript\": \"^5.9.3\"\n  }\n}\n" },
-  { path: "sdk.ts", content: "// The platform's public types surface (the future `@iterate-com/sdk`\n// package), snapshotted into this repo when it was seeded. Import types from\n// it; treat it as read-only — the platform's own copy is the design of\n// record, kept in sync there by lint codegen.\n// \u0063odegen:start {preset: copy, source: ../src/types.ts}\n/**\n * Public ITX capability contract.\n *\n * `/api` — os' one API — gives callers one unauthenticated object.\n * Authentication returns a root catalog, and every object reachable from that\n * catalog is a Cap'n Web / Workers RPC capability. Projects and agents expose\n * stable built-ins (`streams`, `repos`, `workers`, etc.) plus dynamic dotted\n * capabilities mounted on capability hosts (`itx.capabilityHost`,\n * `itx.capabilityHosts.get(path)`). Streams are the durable coordination layer\n * underneath those surfaces: processors, project bootstrap, repo bootstrap,\n * and agent loops all communicate by appending and reducing events.\n */\n\n// -----------------------------------------------------------------------------\n// The four nouns. Keeping them distinct is what makes this system legible:\n//\n// - a SESSION is what `os.authenticate()` returns. It is a catalog that *vends*\n//   itxs; it is not itself an itx.\n// - a PROJECT is the tenant / isolation boundary — a `prj_…` id, its Durable\n//   Objects, its streams. You never hold a \"project object\"; you hold an itx\n//   scoped into a project.\n// - an ITX is a capability context scoped into one project at one path. \"itx\"\n//   is a naming CONVENTION, not a class: an itx is normally an instance of\n//   ProjectRpcTarget whose capability host sits at \"/\" — and sometimes at\n//   \"/agents/…\", which is what \"an agent context\" means. It is the `itx` in\n//   every `async (itx) => { … }` script and what `env.ITX.get()` returns;\n//   `session.projects.get(id)` gives you the itx at the project root.\n// - a CAPABILITY HOST is the durable dynamic-capability table (and script\n//   journal) at one scope path. Each itx fronts exactly one host\n//   (`itx.capabilityHost`; `itx.provideCapability`/`revokeCapability` are\n//   shortcuts onto it); `itx.capabilityHosts.get(path)` addresses any other\n//   scope's host, including the project root at `\"/\"`.\n// -----------------------------------------------------------------------------\n\n/**\n * Self-description of one node in the capability tree — what `__describe()`\n * returns, on EVERY node, by convention.\n *\n * This is the discovery groundwork: `instructions` and `types` are always\n * present and describe THIS node; `children` is the high-level map (one-line\n * blip per child member) so a caller can see what exists without walking;\n * `parent` says where the node sits. Deep discovery is a walk: call\n * `__describe()` on the children you care about. Nodes add structured extras\n * (a project adds `projectId`, an agent adds `whoami`, a capability host adds\n * `capabilities`), so `__describe` is also the identity query — there is no\n * separate `describe()`/`whoami()`.\n *\n * Today each node hand-writes its description (the `describeNode` helper only\n * enforces the shape); the plan is to grow this into a transitive mechanism\n * where a parent composes its children's descriptions.\n */\nexport type Description = {\n  /** Prose: what this node is and how to use it. */\n  instructions: string;\n  /** TypeScript source for this node's surface (\"\" when not yet written). */\n  types: string;\n  /** Discovery map: one-line blip per child member. */\n  children: Record<string, string>;\n  /** Where this node sits / how you reached it. */\n  parent?: string;\n};\n\n/**\n * Every node in the capability tree answers `__describe()` — see\n * {@link Description}. Interfaces below extend this instead of redeclaring;\n * nodes with structured extras (Session, Agent, CapabilityHost, the project\n * itx) declare their own narrowed signature.\n */\nexport interface Describable {\n  __describe(): Promise<Description>;\n}\n\n/**\n * Entry point exposed before any principal or project authority is known.\n *\n * `/api` hands every caller one of these; the only thing it can do is\n * `authenticate(...)`, which on success returns a {@link Session}. This is the\n * canonical Cap'n Web pattern: authority cannot be forged, only handed back by a\n * method that already checked you.\n */\nexport interface UnauthenticatedOs extends Describable {\n  authenticate(input: ItxAuthCredentials): Promise<Session>;\n}\n\n/**\n * What you authenticate into: a catalog that vends itxs.\n *\n * A session is NOT an itx — it is the directory you use to reach one.\n * `projects` is principal-scoped. `streams` and `repos` here are the\n * deployment-wide surfaces backed by `projectId: null`, so only admin/internal\n * auth can reach them.\n */\nexport interface Session {\n  /** Includes `principal` — who this session is (subsumes the old whoami()). */\n  __describe(): Promise<Description & { principal: string }>;\n  projects: ProjectCollection;\n  repos: RepoCollection;\n  streams: StreamCollection;\n}\n\n/** Catalog of projects reachable from a {@link Session}. */\nexport interface ProjectCollection extends Describable {\n  get(projectId: string): Promise<ProjectRpcTarget>;\n  /**\n   * Register and bootstrap a project. By default this resolves once the\n   * bootstrap saga has committed `project/created` — convenient for scripts\n   * and tests that use the project immediately. Pass\n   * `waitUntilCreated: false` to resolve as soon as the project EXISTS\n   * (identity registered, directory primed, bootstrap events appended): the\n   * saga then runs behind the returned handle, and its progress is ordinary\n   * live processor state (`itx.processor.onStateChange` — `state.created`\n   * flips when bootstrap lands). The dashboard uses the fast path to redirect\n   * into the project instantly and play creation progress from pushes.\n   */\n  create(args: {\n    organizationSlug?: string;\n    projectId?: string;\n    slug: string;\n    waitUntilCreated?: boolean;\n  }): Promise<ProjectRpcTarget>;\n  /**\n   * The session's projects, enriched engine-side. Scope is explicit:\n   * - \"mine\" (default for user principals): the caller's own claims (plus\n   *   anything the live context was widened to after a create) — even when\n   *   the socket also carries admin credentials.\n   * - \"deployment\": every deployment-known project from the project\n   *   directory; requires an admin principal (admin secret/cookie or an\n   *   admin-role user). Default for non-user admin principals, which have\n   *   no claims of their own.\n   * Each entry carries the project's {@link ProjectDeploymentStatus} so\n   * callers never probe per project.\n   */\n  list(input?: { scope?: \"mine\" | \"deployment\" }): Promise<ProjectListEntry[]>;\n}\n\n/**\n * Whether a project the directory knows about actually exists in THIS\n * deployment's engine:\n * - `ready` — the project stream's bootstrap saga ran (`state.created`).\n * - `missing` — the engine has no state for it (e.g. the deployment was reset\n *   while the auth worker kept its rows); it can be set up again.\n * - `unknown` — the probe failed (engine hiccup); don't block the list on it.\n */\nexport type ProjectDeploymentStatus = \"ready\" | \"missing\" | \"unknown\";\n\n/** One entry of {@link ProjectCollection.list}. */\nexport type ProjectListEntry = {\n  customDomains: ProjectCustomDomain[];\n  id: string;\n  slug: string;\n  organizationId: string | null;\n  organizationName: string | null;\n  organizationSlug: string | null;\n  deploymentStatus: ProjectDeploymentStatus;\n};\n\n/**\n * An **itx**: a capability context scoped into one project at one path.\n *\n * The same interface serves the project root (path `\"/\"`) and every nested\n * scope (`/agents/bla`, `/agents/slack/ts-124`, …). A nested scope is not a\n * different type — it exposes all the project built-ins below PLUS the dynamic\n * capabilities mounted on its own scope and inherited from every enclosing scope\n * (resolution chains child → parent → project; a nearer scope shadows a farther\n * one). `agent`/`chat` are present only when the scope sits under `/agents/`.\n *\n * DESIGN NOTE — why the built-ins are explicit members, not dynamic entries:\n * this object is an RpcTarget that sits *in front of* the capability-host\n * Durable Object. A call like `itx.streams.get(\"/x\")` resolves against these\n * known members in the isolate and never touches the DO, so the hot path pays\n * no extra round trip to check whether `streams` was shadowed. The deliberate trade-off is that a\n * dynamic capability therefore CANNOT shadow a built-in name — the built-in\n * always wins. If shadowable built-ins turn out to be needed often, we'd move\n * resolution behind the DO and accept the round trip; today we don't.\n */\nexport interface ProjectRpcTarget {\n  /**\n   * Identity + full capability inventory (subsumes the old describe()):\n   * `projectId`/`name`, every reachable capability, the children map, and the\n   * full public type surface in `types`.\n   */\n  __describe(): Promise<ProjectDescription>;\n  ai: Ai;\n  agents: AgentCollection;\n  browser: CfBrowserCapability;\n  /**\n   * This scope's own capability host: the durable capability table behind\n   * this itx (`provideCapability`, `revokeCapability`, `runScript`,\n   * `__describe`). Dynamic dotted calls (`itx.foo.bar(...)`) fall back to it.\n   */\n  capabilityHost: CapabilityHost;\n  /**\n   * Capability hosts of OTHER scopes, by path. `capabilityHosts.get(\"/\")` is\n   * the project root — providing there makes a capability visible to every\n   * scope in the project (child scopes inherit ancestors' mounts).\n   */\n  capabilityHosts: CapabilityHostCollection;\n  /** Shortcut for `capabilityHost.provideCapability` (mounts on THIS scope). */\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  /** Shortcut for `capabilityHost.revokeCapability`. */\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  /** Formatted dashboard/debug info for this itx scope, suitable for Slack messages. */\n  debug(): Promise<string>;\n  egress: ProjectEgress;\n  /** First-party outbound email from the project's own address (`email.send({ to, subject, text })`). */\n  email: EmailCapability;\n  /** Read-only catalogue of known-good itx script snippets (`list()`, `get({ id })`). */\n  examples: ItxExampleCatalog;\n  /** Slack/Google connections + the connection-scoped API proxies (`integrations.gmail`, `integrations.slack`). */\n  integrations: ProjectIntegrations;\n  mcp: McpClientCollection;\n  openapi: OpenApiCollection;\n  processor: StreamProcessorRpc<ProjectProcessorState>;\n  repo: Repo;\n  repos: ProjectRepoCollection;\n  /** Path-addressed sandboxes (`itx.sandboxes.get(path)`) — see {@link SandboxCollection}. */\n  sandboxes: SandboxCollection;\n  secrets: SecretCollection;\n  streams: ProjectStreamCollection;\n  worker: ProjectWorker;\n  workers: DynamicWorkerCollection;\n  // Present only on an agent-scoped itx (path under `/agents/`). `agent` is this\n  // agent's own control surface; `chat` is its web-chat door. They are getters\n  // derived from the scope path, not mounted capabilities — see rpc-targets.ts.\n  agent?: Agent;\n  chat?: AgentChat;\n  /**\n   * THIS agent's own sandbox — the sandbox at the agent's path under the\n   * sandbox prefix (`/agents/bla` → `/sandboxes/cloudflare/agents/bla`).\n   * NOT a built-in: it is a durable itx-expression capability mounted\n   * on the agent's capability host at birth\n   * (`expression: [\"sandboxes\", [\"get\", \"/sandboxes/cloudflare\" + <agent path>]]`), so every\n   * `itx.sandbox.exec(...)` re-resolves through `itx.sandboxes.get` at call\n   * time and dispatches like any provided capability. Created with the agent;\n   * the container boots on the first command and sleeps after idle. Prefer\n   * dotted calls (`await itx.sandbox.exec(\"...\")`) over grabbing the value.\n   */\n  sandbox?: CloudflareSandbox;\n}\n\n/** Agent-local web chat response tool exposed inside agent script execution. */\nexport interface AgentChat extends Describable {\n  sendMessage(input: { message: string }): Promise<StreamEvent>;\n}\n\n/** Workers AI binding exposed through ITX as a project/agent capability. */\nexport interface Ai extends Describable {\n  models(): Promise<unknown>;\n  run(model: string, body: unknown): Promise<unknown>;\n  toMarkdown(\n    ...args: CfMarkdownConversionArgs\n  ): Promise<\n    CfMarkdownSupportedFormat[] | CfMarkdownConversionResult | CfMarkdownConversionResult[]\n  >;\n}\n\nexport type CfMarkdownDocument = {\n  /** Filename including the extension; Cloudflare uses it to choose the converter. */\n  name: string;\n  blob: Blob;\n};\n\nexport type CfMarkdownConversionOptions = {\n  conversionOptions?: {\n    html?: {\n      cssSelector?: string;\n      hostname?: string;\n    };\n    image?: {\n      descriptionLanguage?: string;\n    };\n    pdf?: {\n      excludeMetadata?: boolean;\n    };\n  };\n};\n\nexport type CfMarkdownConversionResult = {\n  name: string;\n  format: \"markdown\" | \"error\";\n  mimeType?: string;\n  tokens?: number;\n  data?: string;\n  error?: string;\n};\n\nexport type CfMarkdownSupportedFormat = {\n  extension: string;\n  mimeType: string;\n};\n\nexport type CfMarkdownConversionArgs =\n  | []\n  | [documents: CfMarkdownDocument | CfMarkdownDocument[], options?: CfMarkdownConversionOptions];\n\nexport type CfBrowserQuickAction =\n  | \"content\"\n  | \"screenshot\"\n  | \"pdf\"\n  | \"markdown\"\n  | \"snapshot\"\n  | \"scrape\"\n  | \"json\"\n  | \"links\"\n  | \"crawl\";\n\nexport type CfBrowserQuickActionOptions = Record<string, unknown> &\n  ({ url: string } | { html: string });\n\n/** Cloudflare Browser Run binding exposed through ITX. */\nexport interface CfBrowserCapability extends Describable {\n  /** Raw Browser Run fetch, primarily for libraries that connect over CDP. */\n  fetch(input: RequestInfo | URL, init?: RequestInit): Promise<Response>;\n  /** Browser Run Quick Actions: content, screenshot, pdf, markdown, snapshot, scrape, json, links, crawl. */\n  quickAction(\n    action: CfBrowserQuickAction,\n    options: CfBrowserQuickActionOptions,\n  ): Promise<Response>;\n}\n\nexport type CfImageTransformOptions = Record<string, unknown>;\n\nexport type CfImageOutputOptions = { format: string } & Record<string, unknown>;\n\nexport type CfImageDrawOptions = Record<string, unknown>;\n\nexport type CfImageTransformInput = {\n  image: ReadableStream<Uint8Array>;\n  transforms?: CfImageTransformOptions[];\n  draws?: Array<{\n    image: ReadableStream<Uint8Array>;\n    options?: CfImageDrawOptions;\n    transforms?: CfImageTransformOptions[];\n  }>;\n  output: CfImageOutputOptions;\n};\n\n/** Cloudflare Images binding exposed through ITX as one-call helpers. */\nexport interface CfImagesCapability extends Describable {\n  info(image: ReadableStream<Uint8Array>): Promise<unknown>;\n  transform(input: CfImageTransformInput): Promise<Response>;\n}\n\nexport type CfVideoTransformOptions = Record<string, unknown>;\n\nexport type CfVideoOutputOptions = {\n  mode: \"video\" | \"spritesheet\" | \"frame\" | \"audio\";\n} & Record<string, unknown>;\n\nexport type CfVideoTransformInput = {\n  video: ReadableStream<Uint8Array>;\n  transform?: CfVideoTransformOptions;\n  output: CfVideoOutputOptions;\n};\n\n/** Cloudflare Media Transformations binding exposed through ITX as one-call helpers. */\nexport interface CfVideosCapability extends Describable {\n  transform(input: CfVideoTransformInput): Promise<Response>;\n}\n\n/** Grouped first-party Cloudflare platform bindings under integrations.cf. */\nexport interface CloudflareIntegrations extends Describable {\n  ai: Ai;\n  browser: CfBrowserCapability;\n  images: CfImagesCapability;\n  videos: CfVideosCapability;\n}\n\n/**\n * First-party outbound email through Cloudflare Email Service. Mail is sent\n * from the project's own address — `<slug>@<project hostname base>`, e.g.\n * `acme@iterate.app` — and an explicit `from` must match it: a project can\n * never send as another project or an arbitrary address. Requires the\n * deployment's sender domain to be onboarded for Email Sending in Cloudflare.\n */\nexport interface EmailCapability extends Describable {\n  send(input: {\n    to: string | string[];\n    subject: string;\n    /** Plain-text body; at least one of text/html is required. */\n    text?: string;\n    html?: string;\n    /** Optional explicit sender; must equal the project's own address. */\n    from?: string;\n  }): Promise<{ from: string; messageId: string | null }>;\n}\n\n// -----------------------------------------------------------------------------\n// Integrations (Slack + Google) — the Phase 12 surface.\n// -----------------------------------------------------------------------------\n\nexport type IntegrationProvider = \"google\" | \"slack\";\n\n/**\n * Slack Web API proxy. `request` is the explicit form; dotted Web API method\n * paths (`itx.integrations.slack.chat.postMessage({...})`) resolve through the dynamic\n * path-call fallback onto `invokeCapability`.\n */\nexport interface SlackCapability extends Describable {\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  request(input: {\n    body?: Record<string, unknown>;\n    method: string;\n  }): Promise<Record<string, unknown>>;\n}\n\nexport type GmailRequestInput = {\n  body?: unknown;\n  headers?: Record<string, string>;\n  method?: string;\n  path: string;\n  query?: Record<string, boolean | number | string | null | undefined>;\n};\n\n/** Gmail REST proxy (`itx.integrations.gmail.request({ path: \"/users/me/messages\" })`). */\nexport interface GmailCapability extends Describable {\n  request(input: GmailRequestInput): Promise<{\n    data: unknown;\n    headers: Record<string, string>;\n    status: number;\n    statusText: string;\n  }>;\n}\n\nexport type IntegrationConnectionStatus = {\n  connected: boolean;\n  displayName: string | null;\n  externalId: string | null;\n  metadata: Record<string, unknown>;\n};\n\nexport type CompleteConnectResult =\n  | { callbackUrl: string | null; ok: true }\n  | { callbackUrl: string | null; error: string; ok: false };\n\n/**\n * Project-scoped integration management (connection status, OAuth, disconnect)\n * plus the connection-scoped API proxies: `integrations.gmail` and\n * `integrations.slack` live here because they only work through the project's\n * connected accounts.\n */\nexport interface ProjectIntegrations extends Describable {\n  /** Cloudflare first-party platform bindings: AI, Browser Run, Images, Media Transformations. */\n  cf: CloudflareIntegrations;\n  /** Gmail REST proxy for the project's connected Google account. */\n  gmail: GmailCapability;\n  /** Slack Web API proxy for the connected workspace (`integrations.slack.chat.postMessage(...)`). */\n  slack: SlackCapability;\n  completeGoogleConnect(input: {\n    code: string;\n    state: string;\n    userId: string | null;\n  }): Promise<CompleteConnectResult>;\n  completeSlackConnect(input: {\n    code: string;\n    state: string;\n    userId: string | null;\n  }): Promise<CompleteConnectResult>;\n  disconnect(input: { provider: IntegrationProvider }): Promise<{ success: true }>;\n  getConnection(input: { provider: IntegrationProvider }): Promise<IntegrationConnectionStatus>;\n  startOAuthFlow(input: {\n    callbackUrl?: string;\n    provider: IntegrationProvider;\n    /** The user to bind the OAuth state to. Browser-supplied, not authority;\n     * the callback's check against the signed state is the backstop. */\n    userId: string;\n  }): Promise<{ authorizationUrl: string }>;\n}\n\nexport type RouteSlackWebhookResult =\n  | { ok: true; projectId: string }\n  | { ignored: \"team-not-claimed\"; ok: true };\n\n/** Agent catalog within one project. */\nexport interface AgentCollection extends Describable {\n  get(path: string): Agent;\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Agent capability surface for message loops and agent-local dynamic tools. */\nexport interface Agent {\n  /** Includes `whoami` (`\"agent <projectId>:<agentPath>\"`), `projectId`, `agentPath`. */\n  __describe(): Promise<Description & { agentPath: string; projectId: string; whoami: string }>;\n  /** The agent scope's own capability host (provide/revoke/runScript/__describe). */\n  capabilityHost: CapabilityHost;\n  /** Shortcut for `capabilityHost.provideCapability` (mounts on THIS agent's scope). */\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  /** Shortcut for `capabilityHost.revokeCapability`. */\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  chat: AgentChat;\n  processor: StreamProcessorRpc<AgentProcessorState>;\n  stream: Stream;\n  sendMessage(message: string): Promise<StreamEvent>;\n  /**\n   * Send-and-wait convenience: appends a user message and resolves with the\n   * agent's next chat reply on this stream. Replies are matched by order, not\n   * correlated per request — concurrent asks on one agent stream interleave\n   * exactly like two people typing into the same chat.\n   */\n  ask(input: {\n    message: string;\n    /** Where the message came from. Defaults to \"web\". */\n    origin?: \"web\" | \"mcp\";\n    /** How long to wait for the reply. Defaults to 45s. */\n    timeoutMs?: number;\n  }): Promise<StreamEvent>;\n}\n\n/** Stream catalog for either a project or the deployment-wide global scope. */\nexport interface StreamCollection extends Describable {\n  get(path: string): Stream;\n}\n\n/** Project-scoped stream catalog with reduced-state listing. */\nexport interface ProjectStreamCollection extends StreamCollection {\n  list(): Promise<StreamListItem[]>;\n}\n\n/**\n * Durable event stream capability.\n *\n * Streams are the public coordination primitive, not an internal queue hidden\n * behind domain methods. Domain helpers can construct common event shapes, but\n * callers and processors still work with explicit events.\n */\nexport interface Stream extends Describable {\n  append(...events: StreamEventInput[]): Promise<StreamEvent[]>;\n  at(path: string): Stream;\n  getEvent(\n    input: { offset: number; idempotencyKey?: never } | { idempotencyKey: string; offset?: never },\n  ): Promise<StreamEvent | undefined>;\n  /** Read one bounded page. Defaults to, and caps at, 500 events. */\n  getEvents(input?: StreamEventReadInput): Promise<StreamEvent[]>;\n  /** Create a disposable pager over a fixed read window. */\n  readEvents(input?: StreamEventReadInput): StreamEventPager;\n  waitForEvent(input: {\n    afterOffset?: number;\n    eventTypes?: readonly string[];\n    predicate?: (event: StreamEvent) => boolean | Promise<boolean>;\n    timeoutMs: number;\n  }): Promise<StreamEvent>;\n  getProcessorRuntimeState(input: {\n    subscriptionKey: string;\n  }): Promise<ProcessorRuntimeState | null>;\n  runtimeState(): Promise<{\n    coreProcessorState: unknown;\n    runtime: {\n      connections: Record<string, unknown>;\n    };\n  }>;\n  subscribe(input: {\n    subscriptionKey?: string;\n    /**\n     * Open the durable configured subscription registered under\n     * `subscriptionKey` (the wake-handshake response) instead of an ephemeral\n     * one. Requires trusted-internal auth and an existing\n     * subscription-configured fact for the key.\n     */\n    configured?: boolean;\n    processEventBatch: ProcessEventBatch;\n    replayAfterOffset?: number;\n    eventTypes?: readonly string[];\n    events?: boolean;\n    subscriber?: unknown;\n  }): Promise<StreamSubscriptionHandle>;\n}\n\n/** Window for bounded stream reads. Offsets are exclusive: `afterOffset: 10` starts at 11. */\nexport type StreamEventReadInput = {\n  /** Exclusive lower bound. Defaults to 0. */\n  afterOffset?: number;\n  /** Exclusive upper bound. Omit/null to read through the current tail. */\n  beforeOffset?: number | null;\n  /** Event types to include. Omit or include \"*\" for all; [] matches none. */\n  eventTypes?: readonly string[];\n  /** Page size, 1-500. Defaults to 500. */\n  limit?: number;\n};\n\n/**\n * Stateful page reader for one stream read window.\n *\n * This is not a live subscription; [] means \"caught up for now\". Dispose it\n * when finished (`using pager = stream.readEvents(...)`).\n */\nexport interface StreamEventPager extends Disposable {\n  /** Returns [] when no newer matching page is currently available. */\n  next(): Promise<StreamEvent[]>;\n}\n\n/** Repo catalog for either a project or the deployment-wide global scope. */\nexport interface RepoCollection extends Describable {\n  create(input: { path: string }): Promise<Repo>;\n  get(path: string): Repo;\n}\n\n/** Project-scoped repo catalog with reduced-state listing. */\nexport interface ProjectRepoCollection extends RepoCollection {\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Git-backed repo capability used by project workers and dynamic worker refs. */\nexport interface Repo extends Describable {\n  commitFiles(input: CommitRepoFilesInput): Promise<CommitRepoFilesResult>;\n  create(): Promise<Repo>;\n  /**\n   * Safely replace text in one committed file and commit the result. The\n   * `oldString` must match exactly once unless `replaceAll` is true.\n   */\n  edit(input: EditRepoFileInput): Promise<EditRepoFileResult>;\n  /** All committed file paths at HEAD. */\n  listFiles(): Promise<{ commitOid: string; paths: string[] }>;\n  processor: StreamProcessorRpc<RepoProcessorState>;\n  /** Committed file contents at HEAD; null when the path does not exist. */\n  readFile(input: { path: string }): Promise<{\n    commitOid: string;\n    content: string;\n    path: string;\n  } | null>;\n  whoami(): Promise<string>;\n}\n\n/**\n * Catalog of sandboxes within one project.\n *\n * A sandbox is addressed by its FULL path, which always lives under\n * `/sandboxes/` — the same domain-prefix convention as `/secrets/...` and\n * `/repos/...`: an agent's sandbox is the agent path under the Cloudflare\n * provider segment (`/sandboxes/cloudflare/agents/...`, exposed as\n * `itx.sandbox` in that agent's scope), and standalone sandboxes\n * conventionally live under `/sandboxes/cloudflare/<anything>`.\n * Getting a sandbox is cheap and does not start a container; the first\n * command does, and the sandbox sleeps again after idle.\n */\nexport interface SandboxCollection extends Describable {\n  get(path: string): Promise<CloudflareSandbox>;\n}\n\n/**\n * One Cloudflare Sandbox: the bare `@cloudflare/sandbox` Durable Object stub,\n * nothing wrapped on top. Whatever the installed SDK exposes is callable —\n * `exec(command)`, `readFile`/`writeFile`/`listFiles`, `startProcess`,\n * `gitCheckout`, `exposePort`, `destroy()`, … — so this contract deliberately\n * does not re-declare that surface (same stance as {@link McpClientRpc}); see\n * https://developers.cloudflare.com/sandbox/ for the API. One addition: the\n * project repo is ALWAYS checked out at `/workspace/repos/project`, which is\n * also the default working directory — a bare `exec(\"ls\")` lists the project.\n * Every command awaits that provisioning internally, so no explicit\n * `ensureProjectRepo()` is needed first (it stays available to await the\n * checkout deterministically).\n */\nexport type CloudflareSandbox = object;\n\n/** Secret catalog within one project. */\nexport interface SecretCollection extends Describable {\n  get(path: string): Secret;\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Path-addressed secret capability. Secret material has no public read API. */\nexport interface Secret extends Describable {\n  describe(): Promise<SecretDescription>;\n  fetch(req: Request): Promise<Response>;\n  processor: StreamProcessorRpc<SecretDescription>;\n  update(input: SecretUpdateInput): Promise<StreamEvent>;\n}\n\nexport type SecretUpdateInput = {\n  egress?: { urls: string[] };\n  material?: string;\n};\n\nexport type SecretDescription = {\n  audit: {\n    lastUsedAt?: string;\n    lastUsedBy?: string;\n    lastUsedUrl?: string;\n    usedCount: number;\n  };\n  egress: { urls: string[] };\n  hasMaterial: boolean;\n};\n\nexport type StreamListItem = {\n  createdAt: string;\n  path: string;\n};\n\nexport type ProjectCustomDomainStatus =\n  | \"requested\"\n  | \"provisioning\"\n  | \"pending_validation\"\n  | \"active\"\n  | \"failed\"\n  | \"removing\";\n\nexport type ProjectCustomDomainValidationRecord = {\n  name: string;\n  status: string | null;\n  value: string;\n};\n\nexport type ProjectCustomDomainDnsRecord = {\n  name: string;\n  value: string;\n};\n\nexport type ProjectCustomDomain = {\n  certificateDelegationCname: ProjectCustomDomainDnsRecord | null;\n  cloudflareHostnameId: string | null;\n  createdAt: string;\n  error: string | null;\n  hostname: string;\n  hostnameStatus: string | null;\n  ownershipVerification: { name: string; value: string } | null;\n  sslStatus: string | null;\n  status: ProjectCustomDomainStatus;\n  updatedAt: string;\n  validationRecords: ProjectCustomDomainValidationRecord[];\n  wildcard: boolean;\n};\n\nexport type ProjectProcessorState = {\n  agents: StreamListItem[];\n  createRequest: { projectId: string; slug: string } | null;\n  created: boolean;\n  customDomains: ProjectCustomDomain[];\n  onboardingActive: boolean;\n  onboardingCompletedAt: string | null;\n  repos: StreamListItem[];\n  secrets: StreamListItem[];\n  streams: StreamListItem[];\n};\n\nexport type AgentProcessorState = {\n  currentRequest:\n    | { phase: \"scheduled\"; requestId: string; scheduledOffset: number }\n    | { phase: \"requested\"; llmRequestId: number }\n    | null;\n  history: Array<{ role: \"user\" | \"assistant\"; content: string }>;\n  llmConfig: { model: string };\n  llmProvider: \"cloudflare-ai\" | \"openai-ws\";\n  pendingTriggerOffset: number | null;\n  inProgressScriptExecutions: Array<{\n    code: string;\n    executionId: string;\n    requestedOffset: number;\n    startedAt: string;\n  }>;\n  scriptExecutionsCompleted: string[];\n  systemPrompt: string;\n};\n\nexport type RepoProcessorState = {\n  artifactName: string | null;\n  created: boolean;\n  defaultBranch: string | null;\n  initialized: boolean;\n  remote: string | null;\n};\n\n// NOTE: there is deliberately no public secret processor state type: a\n// secret's public live state IS its SecretDescription. The internal fold\n// carries the encrypted material; the DO's processor facade projects it away\n// (write-only material) before anything crosses the RPC boundary.\n\nexport type ProcessorSnapshot<State> = {\n  offset: number;\n  state: State;\n};\n\n/**\n * Live handle for one `onStateChange` subscription.\n *\n * `ping()` is the liveness probe: `true` while the subscription is still\n * registered on the live processor, `false` once it was dropped (delivery\n * failure, explicit unsubscribe). The call REJECTS when the hosting Durable\n * Object incarnation is gone. For a subscriber, `false` and a rejection mean\n * the same thing: re-subscribe. Pushes stop silently when a DO restarts or a\n * transport half-opens, so a periodic ping is how a client turns \"silently\n * stale\" into \"detectably dead\".\n */\nexport type ProcessorStateSubscriptionHandle = Disposable & {\n  ping(): boolean | Promise<boolean>;\n  unsubscribe(): void;\n};\n\nexport interface StreamProcessorRpc<State = unknown> {\n  getRuntimeState(): Promise<ProcessorRuntimeState<State>>;\n  /**\n   * Server-push of the processor's reduced state. The callback receives the\n   * durable checkpoint `{ offset, state }` — offset-carrying so clients can\n   * commit pushes and `snapshot()` reads monotonically against each other —\n   * once immediately on subscribe (current state IS the first paint) and then\n   * after every checkpointed batch that changed state.\n   */\n  onStateChange(\n    cb: (snapshot: ProcessorSnapshot<State>) => unknown,\n  ): Promise<ProcessorStateSubscriptionHandle>;\n  snapshot(): Promise<ProcessorSnapshot<State>>;\n  waitUntilEvent(input: { offset: number; timeoutMs?: number }): Promise<void>;\n}\n\n/**\n * Per-stub dispatch options for `DynamicWorkerCollection.get`.\n *\n * `flattenNestedPaths` mirrors `provideCapability`: dotted calls on the stub\n * become ONE `invokeCapability({ path, args })` call that the worker's own\n * `invokeCapability` method dispatches in userspace (one RPC per call),\n * instead of the default member-by-member replay on the entrypoint.\n * `buildBudgetMs` bounds how long a call waits on a cold source build; past\n * it the call fails with an error whose `name` is\n * `\"WorkerBuildInProgressError\"` — the NAME is the contract (it survives\n * Workers RPC; class identity does not), so userspace matches\n * `error.name === \"WorkerBuildInProgressError\"` to render its own building\n * page (the seeded template's router does exactly this).\n */\nexport type DynamicWorkerDispatchOptions = {\n  buildBudgetMs?: number;\n  flattenNestedPaths?: boolean;\n};\n\n/** Capability-tree entry point for ad-hoc project-scoped worker refs. */\nexport interface DynamicWorkerCollection extends Describable {\n  get<T extends object = Record<string, unknown>>(\n    ref: DynamicWorkerRef,\n    options?: DynamicWorkerDispatchOptions,\n  ): DynamicWorkerCapability<T>;\n}\n\n/** Live replacement for project egress. It sees getSecret(...) placeholders, never material. */\nexport type ProjectEgressInterceptor = (req: Request) => Promise<Response>;\n\n/** Disposable handle for one live project egress interception. */\nexport interface ProjectEgressIntercept extends Disposable {\n  release(): Promise<void>;\n}\n\n/**\n * Project-owned egress facet.\n *\n * `fetch` is the explicit outbound door. Dynamic workers' bare `fetch()` uses\n * the same project egress path through the WorkerEntrypoint gateway.\n *\n * `intercept` installs one live runtime replacement on the Project Durable\n * Object. Last writer wins; disposing or releasing the handle clears only the\n * interceptor it installed if it is still current.\n */\nexport interface ProjectEgress extends Describable {\n  fetch(req: Request): Promise<Response>;\n  intercept(handler: ProjectEgressInterceptor): Promise<ProjectEgressIntercept>;\n}\n\n/** What a project itx's `__describe()` returns: the Description convention plus identity and the capability inventory. */\nexport type ProjectDescription = Description & {\n  capabilities: CapabilityDescription[];\n  name: string;\n  projectId: string;\n};\n\nexport type CapabilityDescription = {\n  instructions?: string;\n  path: string[];\n  providedAtOffset?: number;\n  /**\n   * The itx scope path this capability is declared at (`\"/\"`, `\"/agents/bla\"`, …).\n   * Set when a scope reports capabilities it inherited from an enclosing scope,\n   * so the reader can tell a local mount from an inherited one. Absent on\n   * built-ins (they exist at every scope).\n   */\n  scope?: string;\n  type: \"builtin\" | \"live\" | \"itx-expression\";\n  types?: string;\n};\n\nexport interface OpenApiCollection extends Describable {\n  connect(input: OpenApiConnectInput): Promise<OpenApiRpc>;\n}\n\nexport type OpenApiConnectInput = {\n  baseUrl?: string;\n  headers?: Record<string, string>;\n  specUrl: string;\n};\n\nexport type OpenApiRpc = object;\n\nexport interface McpClientCollection extends Describable {\n  connect(input: McpClientConnectInput): Promise<McpClientRpc>;\n  /**\n   * The public Exa MCP server (https://mcp.exa.ai/mcp), pre-connected for every\n   * project: web search and page reading as flat tool calls.\n   * `itx.mcp.exa.web_search_exa({ query, numResults })` searches the web;\n   * `itx.mcp.exa.web_fetch_exa({ urls, maxCharacters })` reads pages as markdown.\n   */\n  exa: McpClientRpc;\n}\n\nexport type McpClientConnectInput = {\n  headers?: Record<string, string>;\n  timeoutMs?: number;\n  url: string;\n};\n\nexport type McpClientRpc = object;\n\n/**\n * Read-only catalogue of known-good itx script snippets — the project-context\n * entries of the REPL \"Examples\" panel. `list()` returns every entry without\n * its code; `get({ id })` returns one entry with the full script body. Agents\n * copy working patterns from here instead of guessing at the surface.\n * Session-level examples (whoami, projects.list) are excluded: they run\n * against the OS Session `authenticate()` returns, which an itx holder does\n * not have.\n */\nexport interface ItxExampleCatalog extends Describable {\n  get(input: { id: string }): Promise<ItxExampleWithCode>;\n  list(): Promise<ItxExampleSummary[]>;\n}\n\nexport type ItxExampleSummary = {\n  description: string;\n  id: string;\n  title: string;\n};\n\nexport type ItxExampleWithCode = ItxExampleSummary & { code: string };\n\n/**\n * The host surface for one capability scope: the durable dynamic-capability\n * table and script journal at one path inside a project (backed by one\n * CapabilityHostDurableObject). Mounting is always local to the host's own\n * scope; reads (`invokeCapability`, `__describe`) chain up through\n * enclosing scopes, so a project-root mount is visible from every scope.\n */\nexport interface CapabilityHost {\n  /** Includes `capabilities`: everything reachable at this scope — own mounts plus inherited ones, tagged with their declaring scope. */\n  __describe(): Promise<Description & { capabilities: CapabilityDescription[]; path: string }>;\n  /** The scope path this host fronts: `\"/\"` is the project root, `/agents/bla` an agent. */\n  path: string;\n  /** Explicit dynamic dispatch; the dotted-path fallback (`itx.foo.bar(...)`) compiles to exactly this call. */\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  runScript(code: string): Promise<{\n    completedEvent: StreamEvent;\n    executionId: string;\n    result: unknown;\n  }>;\n}\n\n/** Capability hosts by scope path within one project (`itx.capabilityHosts`). */\nexport interface CapabilityHostCollection extends Describable {\n  get(path: string): CapabilityHost;\n}\n\n/**\n * Ownership handle for one mounted capability.\n *\n * The offset identifies the exact `capability-provided` event, which lets\n * disposal or explicit revoke remove this mount without racing a newer mount at\n * the same path.\n */\nexport interface CapabilityProvision extends Describable, Disposable {\n  readonly path: string[];\n  readonly providedAtOffset: number;\n  revoke(): Promise<void>;\n}\n\n/** Append input before the stream assigns offset and timestamp. */\nexport type StreamEventSource = {\n  processor?: {\n    slug: string;\n    version: string;\n  };\n  crossPost?: {\n    ruleId: string;\n    from: {\n      createdAt: string;\n      offset: number;\n      path: string;\n      projectId: string | null;\n      type: string;\n    };\n  };\n};\n\nexport type StreamEventInput = {\n  type: string;\n  payload?: Record<string, unknown>;\n  metadata?: Record<string, unknown>;\n  source?: StreamEventSource;\n  idempotencyKey?: string;\n};\n\n/** Durable stream event after commit. */\nexport type StreamEvent = StreamEventInput & {\n  createdAt: string;\n  offset: number;\n};\n\n/** Stable identity for one stream subscription connection. */\nexport type SubscriptionKey = string;\n\n/**\n * Batch delivered to stream processors and live subscribers.\n *\n * Kept named because callback retention, processor hosts, and tests all depend\n * on the same cross-RPC batch envelope.\n */\nexport type StreamEventBatch = {\n  projectId: string | null;\n  path: string;\n  events: StreamEvent[];\n  streamMaxOffset: number;\n  state: unknown;\n};\n\n/**\n * Callback invoked by the stream pump for each delivered batch.\n *\n * It stays as a named type because Workers RPC callback lifecycle helpers need\n * to duplicate, retain, and dispose exactly this callback shape.\n */\nexport type ProcessEventBatch = (batch: StreamEventBatch) => unknown;\n\n/** Serializable snapshot plus optional live runtime debug state for a processor. */\nexport type ProcessorRuntimeState<State = unknown> = {\n  snapshot: { offset: number; state: State };\n  runtime?: Record<string, unknown>;\n};\n\n/**\n * Optional runtime-state callback exposed by a hosted processor.\n *\n * It accepts sync or async implementations because local processors can return\n * immediately, while RPC-backed processors may need an async round trip.\n */\nexport type GetProcessorRuntimeState = () => ProcessorRuntimeState | Promise<ProcessorRuntimeState>;\n\n/**\n * Live subscription handle returned by `Stream.subscribe`.\n *\n * `ping()` mirrors {@link ProcessorStateSubscriptionHandle.ping}: `true` while\n * the connection is still open on the live stream, `false` after it closed\n * (replaced, delivery failure, unsubscribe); it rejects when the stream's\n * Durable Object incarnation is gone. Either non-`true` outcome means the\n * subscriber should re-subscribe.\n */\nexport type StreamSubscriptionHandle = Disposable & {\n  subscriptionKey: SubscriptionKey;\n  streamMaxOffset: number;\n  ping(): boolean | Promise<boolean>;\n  unsubscribe(): void;\n};\n\n/**\n * One repo file mutation.\n *\n * Kept named because public `Repo.commitFiles`, input parsing, and artifact\n * commit implementation all validate the same command shape.\n */\nexport type RepoFileChange =\n  | {\n      path: string;\n      content: string;\n    }\n  | {\n      path: string;\n      delete: true;\n    };\n\n/** Command object for committing a batch of repo file mutations. */\nexport type CommitRepoFilesInput = {\n  author?: { email: string; name: string };\n  branch?: string;\n  changes: RepoFileChange[];\n  message: string;\n};\n\n/** Result returned after a repo commit attempt, including no-op commits. */\nexport type CommitRepoFilesResult = {\n  branch: string;\n  changedPaths: string[];\n  commitOid: string;\n  noChanges: boolean;\n};\n\n/** Command object for a coding-agent-style exact string edit. */\nexport type EditRepoFileInput = {\n  author?: { email: string; name: string };\n  branch?: string;\n  message: string;\n  newString: string;\n  oldString: string;\n  path: string;\n  replaceAll?: boolean;\n};\n\n/** Result returned after an exact string edit commit attempt. */\nexport type EditRepoFileResult = CommitRepoFilesResult & {\n  occurrenceCount: number;\n  path: string;\n};\n\n/** Dynamic invocation envelope used by flattened live capabilities. */\nexport type FlattenedCapabilityInvocation = {\n  args: unknown[];\n  flattenNestedPath?: boolean;\n  path: string[];\n};\n\n/** Target shape for a live capability that wants to receive flattened paths. */\nexport type FlattenedCapabilityTarget = {\n  invokeCapability(input: FlattenedCapabilityInvocation): unknown;\n};\n\n/** Durable expression over the project ITX surface. */\nexport type ItxExpressionStep = string | [method: string, ...args: unknown[]];\nexport type ItxExpression = ItxExpressionStep[];\n\n/** Capability recipe accepted by `provideCapability`. */\nexport type ProvideCapabilityInput =\n  | {\n      capability: unknown;\n      flattenNestedPaths?: false;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      capability: FlattenedCapabilityTarget;\n      flattenNestedPaths: true;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      expression: ItxExpression;\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"itx-expression\";\n      types?: string;\n    };\n\n/** Event payload stored when a capability is mounted on an ITX stream. */\nexport type CapabilityProvidedPayload =\n  | {\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      expression: ItxExpression;\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"itx-expression\";\n      types?: string;\n    };\n\n/** Reduced capability table row: payload plus the providing event offset. */\nexport type CapabilityRecord = CapabilityProvidedPayload & {\n  providedAtOffset: number;\n};\n\n/** Revoke command for the current mount at a path or one exact mount offset. */\nexport type RevokeCapabilityInput = {\n  path: string[];\n  providedAtOffset?: number;\n};\n\n/**\n * Credentials accepted by `UnauthenticatedOs.authenticate`.\n *\n * - `from-server-cookie` — the browser lane: the deployment's admin cookie or\n *   the signed-in user's session cookie riding the WebSocket handshake.\n * - `bearer` — an auth access token presented as RPC data.\n * - `admin-secret` — the deployment admin API secret (CLI / tooling / e2e).\n * - `impersonate` — admin-secret-gated fake principal, for test suites that\n *   exercise per-project confinement without minting real users.\n */\nexport type ItxAuthCredentials =\n  | { type: \"from-server-cookie\" }\n  | { type: \"bearer\"; token: string }\n  | { type: \"admin-secret\"; secret: string }\n  | { type: \"impersonate\"; secret: string; token: ItxAuthToken };\n\n/** Principal shape for `impersonate` credentials. */\nexport type ItxAuthToken =\n  | { type: \"admin\"; principal?: string }\n  | { type: \"user\"; principal: string; projectScopes: string[] };\n\n/** Authority object carried by server-side RPC target instances. */\nexport interface ItxAuth {\n  readonly principal: string;\n  isAdmin(): boolean;\n  canAccessProject(projectId: string): boolean;\n  assertCanAccessProject(projectId: string | null): void;\n  listAccessibleProjects(): string[];\n  /**\n   * Async access check that may consult the project directory (source of\n   * truth) when synchronous claims miss — see the auth adapter. Optional so\n   * in-process trusted contexts stay trivially constructible.\n   */\n  ensureCanAccessProject?(projectId: string): Promise<void>;\n}\n\n/**\n * Where a dynamic worker's source files come from.\n *\n * `inline` supplies the file map directly — the primitive behind run-script and\n * worker-backed provided capabilities where the caller hands over a small\n * TypeScript entry file, helpers, and optionally a `package.json`. `repo` names\n * a project repo snapshot: a branch (late-bound, so future commits affect the\n * next use) or a pinned commit, narrowed by include/exclude glob masks so a\n * large repo does not become build input by default.\n */\nexport type WorkerFileSource =\n  | {\n      type: \"inline\";\n      files: Record<string, string>;\n    }\n  | {\n      type: \"repo\";\n      repoPath: string;\n      /**\n       * Defaults to the repo's default branch when omitted. A pinned commit\n       * may name the branch it lives on — clones are single-branch, so an\n       * off-default-branch commit is unreachable without it.\n       */\n      ref?: { branch: string } | { commitOid: string; branch?: string };\n      include?: string[];\n      exclude?: string[];\n    };\n\n/** Loader names accepted by Cloudflare's worker bundler `loader` option. */\nexport type WorkerBundlerLoader =\n  | \"js\"\n  | \"jsx\"\n  | \"ts\"\n  | \"tsx\"\n  | \"json\"\n  | \"css\"\n  | \"text\"\n  | \"binary\"\n  | \"base64\"\n  | \"dataurl\";\n\n/**\n * Build options for a dynamic worker.\n *\n * This mirrors Cloudflare's `CreateWorkerOptions` from\n * `@cloudflare/worker-bundler` minus `files` (OS supplies files from the\n * selected {@link WorkerFileSource}) — deliberately not a parallel option\n * language (drift fails typecheck via the assignability pin in\n * domains/workers/schemas.ts; this file stays import-free because it is\n * published verbatim as the itx types surface). `bundle: false` is allowed; the invariant is one OS\n * materialization pipeline, not one bundled output file. When the file map has\n * a `package.json` with dependencies, the bundler installs them from the npm\n * registry at build time.\n */\nexport type WorkerBuildOptions = {\n  /** Entry point file path relative to the source root (e.g. \"worker.ts\"). */\n  entryPoint?: string;\n  /** Bundle all dependencies into a single output file. Default: true. */\n  bundle?: boolean;\n  /** Modules kept external (\"cloudflare:*\" always is). */\n  externals?: string[];\n  /** Target environment. Default: \"es2022\". */\n  target?: string;\n  minify?: boolean;\n  sourcemap?: boolean;\n  /** npm registry URL for dependency installs. */\n  registry?: string;\n  jsx?: \"transform\" | \"preserve\" | \"automatic\";\n  jsxImportSource?: string;\n  define?: Record<string, string>;\n  loader?: Record<string, WorkerBundlerLoader>;\n  conditions?: string[];\n  virtualModules?: Record<string, string>;\n};\n\n/**\n * Declarative source for a dynamic worker: an orthogonal file source plus\n * Cloudflare-compatible build options.\n *\n * Materialization resolves `files` to a file map and builds it through\n * Cloudflare's worker bundler; the loader-ready output is cached by a\n * deterministic build key, so the same source+options never builds twice.\n */\nexport type DynamicWorkerSource = {\n  files: WorkerFileSource;\n  options?: WorkerBuildOptions;\n};\n\ntype DynamicWorkerRefBase = {\n  /**\n   * ITX scope path for the worker's `env.ITX` binding and for stateful worker\n   * Durable Object names. This is intentionally not the mounted capability path:\n   * one worker can be mounted at `db`, `counter`, etc. while all events still\n   * belong to the host stream path.\n   */\n  path: string;\n  source: DynamicWorkerSource;\n};\n\n/**\n * Stateless workers are WorkerEntrypoint exports loaded directly from source.\n *\n * `props` are passed to `worker.getEntrypoint(name, { props })` and appear as\n * `this.ctx.props` inside the exported WorkerEntrypoint. They deliberately live\n * only on stateless refs: Durable Object facets are started with\n * `ctx.facets.get(name, () => ({ class, id? }))`, which does not accept\n * WorkerEntrypoint-style props.\n */\nexport type StatelessDynamicWorkerRef = DynamicWorkerRefBase & {\n  type: \"stateless\";\n  entrypoint?: string;\n  props?: Record<string, JsonValue>;\n};\n\n/**\n * Stateful workers are Durable Object class exports hosted by\n * `StatefulWorkerDurableObject`.\n *\n * `durableWorkerKey` is the durable identity under `{ projectId, path }`. It is\n * not a source cache key: source changes deliberately affect the next use of the\n * same durable worker identity.\n */\nexport type StatefulDynamicWorkerRef = DynamicWorkerRefBase & {\n  type: \"stateful\";\n  className: string;\n  durableWorkerKey: string;\n  /**\n   * What a call does when the worker's source changed since the running\n   * version. `\"block\"` (default) waits for the rebuild — commit-then-call\n   * sees the new code. `\"stale-while-rebuild\"` keeps answering with the\n   * running version and swaps to the new build in the background: better\n   * availability, but the next few calls after a commit may see old code.\n   * The policy rides the REF, not the durable identity — callers sharing one\n   * `durableWorkerKey` should agree on it (and on `source`), or each call\n   * flips the facet to its own version.\n   */\n  updatePolicy?: \"block\" | \"stale-while-rebuild\";\n};\n\n/** Worker recipe accepted by `workers.get` and worker-backed capabilities. */\nexport type DynamicWorkerRef = StatelessDynamicWorkerRef | StatefulDynamicWorkerRef;\n\n/** Dynamic worker RPC stub plus the disposal operation owned by the caller. */\nexport type DynamicWorkerCapability<T extends object = Record<string, unknown>> = T & Disposable;\n\n/**\n * Slack Web API surface exposed by the seeded project worker\n * (`itx.worker.slack.chat.postMessage({...})`).\n *\n * The seeded repo implements this in userland with the real `@slack/web-api`\n * package (installed by the worker build pipeline from its `package.json`), so\n * any nested Web API method family resolves — the index signature reflects\n * that this tree is as wide as the SDK's.\n */\nexport interface ProjectWorkerSlack {\n  chat: {\n    postMessage(input: Record<string, unknown>): Promise<Record<string, unknown>>;\n  } & Record<string, unknown>;\n  [family: string]: unknown;\n}\n\n/**\n * Default seeded project worker contract.\n *\n * This documents the reference repo's `worker.ts` only. Arbitrary dynamic\n * workers should be typed by callers through `workers.get<T>(ref)`. The\n * platform dispatches to it with flattened paths, so the worker implements\n * `invokeCapability` in userspace and every dotted call — including any\n * nested `slack.*` Web API family — is one RPC.\n */\nexport interface ProjectWorker {\n  fetch(req: Request): Promise<Response>;\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  processEvent(input: { event: StreamEvent }): Promise<void>;\n  slack: ProjectWorkerSlack;\n}\n\n/** JSON subset accepted by WorkerEntrypoint props and script results. */\nexport type JsonValue =\n  | string\n  | number\n  | boolean\n  | null\n  | JsonValue[]\n  | { [key: string]: JsonValue };\n\n/**\n * Minimal ExecutionContext shape the RPC adapter layer needs.\n *\n * This stays at the bottom because it is server-side plumbing, not part of the\n * client-facing ITX contract. It is exported only so domain hosts can inject\n * project/agent capability targets without importing the full worker module.\n */\nexport type CfExecutionContext = {\n  exports: ExecutionContext[\"exports\"];\n  /** Required, not optional: budget-expired dynamic worker builds are handed\n   * to it (worker-loader.ts withBuildBudget) — a silent no-op here would\n   * strand every cold build the caller gave up on. Hosts without a real\n   * runtime hook must still supply an explicit function. */\n  waitUntil: ExecutionContext[\"waitUntil\"];\n};\n// \u0063odegen:end\n" },
+  { path: "sdk.ts", content: "// The platform's public types surface (the future `@iterate-com/sdk`\n// package), snapshotted into this repo when it was seeded. Import types from\n// it; treat it as read-only — the platform's own copy is the design of\n// record, kept in sync there by lint codegen.\n// \u0063odegen:start {preset: copy, source: ../src/types.ts}\n/**\n * Public ITX capability contract.\n *\n * `/api` — os' one API — gives callers one unauthenticated object.\n * Authentication returns a root catalog, and every object reachable from that\n * catalog is a Cap'n Web / Workers RPC capability. Projects and agents expose\n * stable built-ins (`streams`, `repos`, `workers`, etc.) plus dynamic dotted\n * capabilities mounted on capability hosts (`itx.capabilityHost`,\n * `itx.capabilityHosts.get(path)`). Streams are the durable coordination layer\n * underneath those surfaces: processors, project bootstrap, repo bootstrap,\n * and agent loops all communicate by appending and reducing events.\n */\n\n// -----------------------------------------------------------------------------\n// The four nouns. Keeping them distinct is what makes this system legible:\n//\n// - a SESSION is what `os.authenticate()` returns. It is a catalog that *vends*\n//   itxs; it is not itself an itx.\n// - a PROJECT is the tenant / isolation boundary — a `prj_…` id, its Durable\n//   Objects, its streams. You never hold a \"project object\"; you hold an itx\n//   scoped into a project.\n// - an ITX is a capability context scoped into one project at one path. \"itx\"\n//   is a naming CONVENTION, not a class: an itx is normally an instance of\n//   ProjectRpcTarget whose capability host sits at \"/\" — and sometimes at\n//   \"/agents/…\", which is what \"an agent context\" means. It is the `itx` in\n//   every `async (itx) => { … }` script and what `env.ITX.get()` returns;\n//   `session.projects.get(id)` gives you the itx at the project root.\n// - a CAPABILITY HOST is the durable dynamic-capability table (and script\n//   journal) at one scope path. Each itx fronts exactly one host\n//   (`itx.capabilityHost`; `itx.provideCapability`/`revokeCapability` are\n//   shortcuts onto it); `itx.capabilityHosts.get(path)` addresses any other\n//   scope's host, including the project root at `\"/\"`.\n// -----------------------------------------------------------------------------\n\n/**\n * Self-description of one node in the capability tree — what `__describe()`\n * returns, on EVERY node, by convention.\n *\n * This is the discovery groundwork: `instructions` and `types` are always\n * present and describe THIS node; `children` is the high-level map (one-line\n * blip per child member) so a caller can see what exists without walking;\n * `parent` says where the node sits. Deep discovery is a walk: call\n * `__describe()` on the children you care about. Nodes add structured extras\n * (a project adds `projectId`, an agent adds `whoami`, a capability host adds\n * `capabilities`), so `__describe` is also the identity query — there is no\n * separate `describe()`/`whoami()`.\n *\n * Today each node hand-writes its description (the `describeNode` helper only\n * enforces the shape); the plan is to grow this into a transitive mechanism\n * where a parent composes its children's descriptions.\n */\nexport type Description = {\n  /** Prose: what this node is and how to use it. */\n  instructions: string;\n  /** TypeScript source for this node's surface (\"\" when not yet written). */\n  types: string;\n  /** Discovery map: one-line blip per child member. */\n  children: Record<string, string>;\n  /** Where this node sits / how you reached it. */\n  parent?: string;\n};\n\n/**\n * Every node in the capability tree answers `__describe()` — see\n * {@link Description}. Interfaces below extend this instead of redeclaring;\n * nodes with structured extras (Session, Agent, CapabilityHost, the project\n * itx) declare their own narrowed signature.\n */\nexport interface Describable {\n  __describe(): Promise<Description>;\n}\n\n/**\n * Entry point exposed before any principal or project authority is known.\n *\n * `/api` hands every caller one of these; the only thing it can do is\n * `authenticate(...)`, which on success returns a {@link Session}. This is the\n * canonical Cap'n Web pattern: authority cannot be forged, only handed back by a\n * method that already checked you.\n */\nexport interface UnauthenticatedOs extends Describable {\n  authenticate(input: ItxAuthCredentials): Promise<Session>;\n}\n\n/**\n * What you authenticate into: a catalog that vends itxs.\n *\n * A session is NOT an itx — it is the directory you use to reach one.\n * `projects` is principal-scoped. `streams` and `repos` here are the\n * deployment-wide surfaces backed by `projectId: null`, so only admin/internal\n * auth can reach them.\n */\nexport interface Session {\n  /** Includes `principal` — who this session is (subsumes the old whoami()). */\n  __describe(): Promise<Description & { principal: string }>;\n  projects: ProjectCollection;\n  repos: RepoCollection;\n  streams: StreamCollection;\n}\n\n/** Catalog of projects reachable from a {@link Session}. */\nexport interface ProjectCollection extends Describable {\n  get(projectId: string): Promise<ProjectRpcTarget>;\n  /**\n   * Register and bootstrap a project. By default this resolves once the\n   * bootstrap saga has committed `project/created` — convenient for scripts\n   * and tests that use the project immediately. Pass\n   * `waitUntilCreated: false` to resolve as soon as the project EXISTS\n   * (identity registered, directory primed, bootstrap events appended): the\n   * saga then runs behind the returned handle, and its progress is ordinary\n   * live processor state (`itx.processor.onStateChange` — `state.created`\n   * flips when bootstrap lands). The dashboard uses the fast path to redirect\n   * into the project instantly and play creation progress from pushes.\n   */\n  create(args: {\n    organizationSlug?: string;\n    projectId?: string;\n    slug: string;\n    waitUntilCreated?: boolean;\n  }): Promise<ProjectRpcTarget>;\n  /**\n   * The session's projects, enriched engine-side. Scope is explicit:\n   * - \"mine\" (default for user principals): the caller's own claims (plus\n   *   anything the live context was widened to after a create) — even when\n   *   the socket also carries admin credentials.\n   * - \"deployment\": every deployment-known project from the project\n   *   directory; requires an admin principal (admin secret/cookie or an\n   *   admin-role user). Default for non-user admin principals, which have\n   *   no claims of their own.\n   * Each entry carries the project's {@link ProjectDeploymentStatus} so\n   * callers never probe per project.\n   */\n  list(input?: { scope?: \"mine\" | \"deployment\" }): Promise<ProjectListEntry[]>;\n}\n\n/**\n * Whether a project the directory knows about actually exists in THIS\n * deployment's engine:\n * - `ready` — the project stream's bootstrap saga ran (`state.created`).\n * - `missing` — the engine has no state for it (e.g. the deployment was reset\n *   while the auth worker kept its rows); it can be set up again.\n * - `unknown` — the probe failed (engine hiccup); don't block the list on it.\n */\nexport type ProjectDeploymentStatus = \"ready\" | \"missing\" | \"unknown\";\n\n/** One entry of {@link ProjectCollection.list}. */\nexport type ProjectListEntry = {\n  id: string;\n  slug: string;\n  organizationId: string | null;\n  organizationName: string | null;\n  organizationSlug: string | null;\n  deploymentStatus: ProjectDeploymentStatus;\n};\n\n/**\n * An **itx**: a capability context scoped into one project at one path.\n *\n * The same interface serves the project root (path `\"/\"`) and every nested\n * scope (`/agents/bla`, `/agents/slack/ts-124`, …). A nested scope is not a\n * different type — it exposes all the project built-ins below PLUS the dynamic\n * capabilities mounted on its own scope and inherited from every enclosing scope\n * (resolution chains child → parent → project; a nearer scope shadows a farther\n * one). `agent`/`chat` are present only when the scope sits under `/agents/`.\n *\n * DESIGN NOTE — why the built-ins are explicit members, not dynamic entries:\n * this object is an RpcTarget that sits *in front of* the capability-host\n * Durable Object. A call like `itx.streams.get(\"/x\")` resolves against these\n * known members in the isolate and never touches the DO, so the hot path pays\n * no extra round trip to check whether `streams` was shadowed. The deliberate trade-off is that a\n * dynamic capability therefore CANNOT shadow a built-in name — the built-in\n * always wins. If shadowable built-ins turn out to be needed often, we'd move\n * resolution behind the DO and accept the round trip; today we don't.\n */\nexport interface ProjectRpcTarget {\n  /**\n   * Identity + full capability inventory (subsumes the old describe()):\n   * `projectId`/`name`, every reachable capability, the children map, and the\n   * full public type surface in `types`.\n   */\n  __describe(): Promise<ProjectDescription>;\n  ai: Ai;\n  agents: AgentCollection;\n  browser: CfBrowserCapability;\n  /**\n   * This scope's own capability host: the durable capability table behind\n   * this itx (`provideCapability`, `revokeCapability`, `runScript`,\n   * `__describe`). Dynamic dotted calls (`itx.foo.bar(...)`) fall back to it.\n   */\n  capabilityHost: CapabilityHost;\n  /**\n   * Capability hosts of OTHER scopes, by path. `capabilityHosts.get(\"/\")` is\n   * the project root — providing there makes a capability visible to every\n   * scope in the project (child scopes inherit ancestors' mounts).\n   */\n  capabilityHosts: CapabilityHostCollection;\n  /** Shortcut for `capabilityHost.provideCapability` (mounts on THIS scope). */\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  /** Shortcut for `capabilityHost.revokeCapability`. */\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  /** Formatted dashboard/debug info for this itx scope, suitable for Slack messages. */\n  debug(): Promise<string>;\n  egress: ProjectEgress;\n  /** First-party outbound email from the project's own address (`email.send({ to, subject, text })`). */\n  email: EmailCapability;\n  /** Read-only catalogue of known-good itx script snippets (`list()`, `get({ id })`). */\n  examples: ItxExampleCatalog;\n  /** Slack/Google connections + the connection-scoped API proxies (`integrations.gmail`, `integrations.slack`). */\n  integrations: ProjectIntegrations;\n  mcp: McpClientCollection;\n  openapi: OpenApiCollection;\n  processor: StreamProcessorRpc<ProjectProcessorState>;\n  repo: Repo;\n  repos: ProjectRepoCollection;\n  /** Path-addressed sandboxes (`itx.sandboxes.get(path)`) — see {@link SandboxCollection}. */\n  sandboxes: SandboxCollection;\n  secrets: SecretCollection;\n  streams: ProjectStreamCollection;\n  worker: ProjectWorker;\n  workers: DynamicWorkerCollection;\n  // Present only on an agent-scoped itx (path under `/agents/`). `agent` is this\n  // agent's own control surface; `chat` is its web-chat door. They are getters\n  // derived from the scope path, not mounted capabilities — see rpc-targets.ts.\n  agent?: Agent;\n  chat?: AgentChat;\n  /**\n   * THIS agent's own sandbox — the sandbox at the agent's path under the\n   * sandbox prefix (`/agents/bla` → `/sandboxes/cloudflare/agents/bla`).\n   * NOT a built-in: it is a durable itx-expression capability mounted\n   * on the agent's capability host at birth\n   * (`expression: [\"sandboxes\", [\"get\", \"/sandboxes/cloudflare\" + <agent path>]]`), so every\n   * `itx.sandbox.exec(...)` re-resolves through `itx.sandboxes.get` at call\n   * time and dispatches like any provided capability. Created with the agent;\n   * the container boots on the first command and sleeps after idle. Prefer\n   * dotted calls (`await itx.sandbox.exec(\"...\")`) over grabbing the value.\n   */\n  sandbox?: CloudflareSandbox;\n}\n\n/** Agent-local web chat response tool exposed inside agent script execution. */\nexport interface AgentChat extends Describable {\n  sendMessage(input: { message: string }): Promise<StreamEvent>;\n}\n\n/** Workers AI binding exposed through ITX as a project/agent capability. */\nexport interface Ai extends Describable {\n  models(): Promise<unknown>;\n  run(model: string, body: unknown): Promise<unknown>;\n  toMarkdown(\n    ...args: CfMarkdownConversionArgs\n  ): Promise<\n    CfMarkdownSupportedFormat[] | CfMarkdownConversionResult | CfMarkdownConversionResult[]\n  >;\n}\n\nexport type CfMarkdownDocument = {\n  /** Filename including the extension; Cloudflare uses it to choose the converter. */\n  name: string;\n  blob: Blob;\n};\n\nexport type CfMarkdownConversionOptions = {\n  conversionOptions?: {\n    html?: {\n      cssSelector?: string;\n      hostname?: string;\n    };\n    image?: {\n      descriptionLanguage?: string;\n    };\n    pdf?: {\n      excludeMetadata?: boolean;\n    };\n  };\n};\n\nexport type CfMarkdownConversionResult = {\n  name: string;\n  format: \"markdown\" | \"error\";\n  mimeType?: string;\n  tokens?: number;\n  data?: string;\n  error?: string;\n};\n\nexport type CfMarkdownSupportedFormat = {\n  extension: string;\n  mimeType: string;\n};\n\nexport type CfMarkdownConversionArgs =\n  | []\n  | [documents: CfMarkdownDocument | CfMarkdownDocument[], options?: CfMarkdownConversionOptions];\n\nexport type CfBrowserQuickAction =\n  | \"content\"\n  | \"screenshot\"\n  | \"pdf\"\n  | \"markdown\"\n  | \"snapshot\"\n  | \"scrape\"\n  | \"json\"\n  | \"links\"\n  | \"crawl\";\n\nexport type CfBrowserQuickActionOptions = Record<string, unknown> &\n  ({ url: string } | { html: string });\n\n/** Cloudflare Browser Run binding exposed through ITX. */\nexport interface CfBrowserCapability extends Describable {\n  /** Raw Browser Run fetch, primarily for libraries that connect over CDP. */\n  fetch(input: RequestInfo | URL, init?: RequestInit): Promise<Response>;\n  /** Browser Run Quick Actions: content, screenshot, pdf, markdown, snapshot, scrape, json, links, crawl. */\n  quickAction(\n    action: CfBrowserQuickAction,\n    options: CfBrowserQuickActionOptions,\n  ): Promise<Response>;\n}\n\nexport type CfImageTransformOptions = Record<string, unknown>;\n\nexport type CfImageOutputOptions = { format: string } & Record<string, unknown>;\n\nexport type CfImageDrawOptions = Record<string, unknown>;\n\nexport type CfImageTransformInput = {\n  image: ReadableStream<Uint8Array>;\n  transforms?: CfImageTransformOptions[];\n  draws?: Array<{\n    image: ReadableStream<Uint8Array>;\n    options?: CfImageDrawOptions;\n    transforms?: CfImageTransformOptions[];\n  }>;\n  output: CfImageOutputOptions;\n};\n\n/** Cloudflare Images binding exposed through ITX as one-call helpers. */\nexport interface CfImagesCapability extends Describable {\n  info(image: ReadableStream<Uint8Array>): Promise<unknown>;\n  transform(input: CfImageTransformInput): Promise<Response>;\n}\n\nexport type CfVideoTransformOptions = Record<string, unknown>;\n\nexport type CfVideoOutputOptions = {\n  mode: \"video\" | \"spritesheet\" | \"frame\" | \"audio\";\n} & Record<string, unknown>;\n\nexport type CfVideoTransformInput = {\n  video: ReadableStream<Uint8Array>;\n  transform?: CfVideoTransformOptions;\n  output: CfVideoOutputOptions;\n};\n\n/** Cloudflare Media Transformations binding exposed through ITX as one-call helpers. */\nexport interface CfVideosCapability extends Describable {\n  transform(input: CfVideoTransformInput): Promise<Response>;\n}\n\n/** Grouped first-party Cloudflare platform bindings under integrations.cf. */\nexport interface CloudflareIntegrations extends Describable {\n  ai: Ai;\n  browser: CfBrowserCapability;\n  images: CfImagesCapability;\n  videos: CfVideosCapability;\n}\n\n/**\n * First-party outbound email through Cloudflare Email Service. Mail is sent\n * from the project's own address — `<slug>@<project hostname base>`, e.g.\n * `acme@iterate.app` — and an explicit `from` must match it: a project can\n * never send as another project or an arbitrary address. Requires the\n * deployment's sender domain to be onboarded for Email Sending in Cloudflare.\n */\nexport interface EmailCapability extends Describable {\n  send(input: {\n    to: string | string[];\n    subject: string;\n    /** Plain-text body; at least one of text/html is required. */\n    text?: string;\n    html?: string;\n    /** Optional explicit sender; must equal the project's own address. */\n    from?: string;\n  }): Promise<{ from: string; messageId: string | null }>;\n}\n\n// -----------------------------------------------------------------------------\n// Integrations (Slack + Google) — the Phase 12 surface.\n// -----------------------------------------------------------------------------\n\nexport type IntegrationProvider = \"google\" | \"slack\";\n\n/**\n * Slack Web API proxy. `request` is the explicit form; dotted Web API method\n * paths (`itx.integrations.slack.chat.postMessage({...})`) resolve through the dynamic\n * path-call fallback onto `invokeCapability`.\n */\nexport interface SlackCapability extends Describable {\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  request(input: {\n    body?: Record<string, unknown>;\n    method: string;\n  }): Promise<Record<string, unknown>>;\n}\n\nexport type GmailRequestInput = {\n  body?: unknown;\n  headers?: Record<string, string>;\n  method?: string;\n  path: string;\n  query?: Record<string, boolean | number | string | null | undefined>;\n};\n\n/** Gmail REST proxy (`itx.integrations.gmail.request({ path: \"/users/me/messages\" })`). */\nexport interface GmailCapability extends Describable {\n  request(input: GmailRequestInput): Promise<{\n    data: unknown;\n    headers: Record<string, string>;\n    status: number;\n    statusText: string;\n  }>;\n}\n\nexport type IntegrationConnectionStatus = {\n  connected: boolean;\n  displayName: string | null;\n  externalId: string | null;\n  metadata: Record<string, unknown>;\n};\n\nexport type CompleteConnectResult =\n  | { callbackUrl: string | null; ok: true }\n  | { callbackUrl: string | null; error: string; ok: false };\n\n/**\n * Project-scoped integration management (connection status, OAuth, disconnect)\n * plus the connection-scoped API proxies: `integrations.gmail` and\n * `integrations.slack` live here because they only work through the project's\n * connected accounts.\n */\nexport interface ProjectIntegrations extends Describable {\n  /** Cloudflare first-party platform bindings: AI, Browser Run, Images, Media Transformations. */\n  cf: CloudflareIntegrations;\n  /** Gmail REST proxy for the project's connected Google account. */\n  gmail: GmailCapability;\n  /** Slack Web API proxy for the connected workspace (`integrations.slack.chat.postMessage(...)`). */\n  slack: SlackCapability;\n  completeGoogleConnect(input: {\n    code: string;\n    state: string;\n    userId: string | null;\n  }): Promise<CompleteConnectResult>;\n  completeSlackConnect(input: {\n    code: string;\n    state: string;\n    userId: string | null;\n  }): Promise<CompleteConnectResult>;\n  disconnect(input: { provider: IntegrationProvider }): Promise<{ success: true }>;\n  getConnection(input: { provider: IntegrationProvider }): Promise<IntegrationConnectionStatus>;\n  startOAuthFlow(input: {\n    callbackUrl?: string;\n    provider: IntegrationProvider;\n    /** The user to bind the OAuth state to. Browser-supplied, not authority;\n     * the callback's check against the signed state is the backstop. */\n    userId: string;\n  }): Promise<{ authorizationUrl: string }>;\n}\n\nexport type RouteSlackWebhookResult =\n  | { ok: true; projectId: string }\n  | { ignored: \"team-not-claimed\"; ok: true };\n\n/** Agent catalog within one project. */\nexport interface AgentCollection extends Describable {\n  get(path: string): Agent;\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Agent capability surface for message loops and agent-local dynamic tools. */\nexport interface Agent {\n  /** Includes `whoami` (`\"agent <projectId>:<agentPath>\"`), `projectId`, `agentPath`. */\n  __describe(): Promise<Description & { agentPath: string; projectId: string; whoami: string }>;\n  /** The agent scope's own capability host (provide/revoke/runScript/__describe). */\n  capabilityHost: CapabilityHost;\n  /** Shortcut for `capabilityHost.provideCapability` (mounts on THIS agent's scope). */\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  /** Shortcut for `capabilityHost.revokeCapability`. */\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  chat: AgentChat;\n  processor: StreamProcessorRpc<AgentProcessorState>;\n  stream: Stream;\n  sendMessage(message: string): Promise<StreamEvent>;\n  /**\n   * Send-and-wait convenience: appends a user message and resolves with the\n   * agent's next chat reply on this stream. Replies are matched by order, not\n   * correlated per request — concurrent asks on one agent stream interleave\n   * exactly like two people typing into the same chat.\n   */\n  ask(input: {\n    message: string;\n    /** Where the message came from. Defaults to \"web\". */\n    origin?: \"web\" | \"mcp\";\n    /** How long to wait for the reply. Defaults to 45s. */\n    timeoutMs?: number;\n  }): Promise<StreamEvent>;\n}\n\n/** Stream catalog for either a project or the deployment-wide global scope. */\nexport interface StreamCollection extends Describable {\n  get(path: string): Stream;\n}\n\n/** Project-scoped stream catalog with reduced-state listing. */\nexport interface ProjectStreamCollection extends StreamCollection {\n  list(): Promise<StreamListItem[]>;\n}\n\n/**\n * Durable event stream capability.\n *\n * Streams are the public coordination primitive, not an internal queue hidden\n * behind domain methods. Domain helpers can construct common event shapes, but\n * callers and processors still work with explicit events.\n */\nexport interface Stream extends Describable {\n  append(...events: StreamEventInput[]): Promise<StreamEvent[]>;\n  at(path: string): Stream;\n  getEvent(\n    input: { offset: number; idempotencyKey?: never } | { idempotencyKey: string; offset?: never },\n  ): Promise<StreamEvent | undefined>;\n  /** Read one bounded page. Defaults to, and caps at, 500 events. */\n  getEvents(input?: StreamEventReadInput): Promise<StreamEvent[]>;\n  /** Create a disposable pager over a fixed read window. */\n  readEvents(input?: StreamEventReadInput): StreamEventPager;\n  waitForEvent(input: {\n    afterOffset?: number;\n    eventTypes?: readonly string[];\n    predicate?: (event: StreamEvent) => boolean | Promise<boolean>;\n    timeoutMs: number;\n  }): Promise<StreamEvent>;\n  getProcessorRuntimeState(input: {\n    subscriptionKey: string;\n  }): Promise<ProcessorRuntimeState | null>;\n  runtimeState(): Promise<{\n    coreProcessorState: unknown;\n    runtime: {\n      connections: Record<string, unknown>;\n    };\n  }>;\n  subscribe(input: {\n    subscriptionKey?: string;\n    /**\n     * Open the durable configured subscription registered under\n     * `subscriptionKey` (the wake-handshake response) instead of an ephemeral\n     * one. Requires trusted-internal auth and an existing\n     * subscription-configured fact for the key.\n     */\n    configured?: boolean;\n    processEventBatch: ProcessEventBatch;\n    replayAfterOffset?: number;\n    eventTypes?: readonly string[];\n    events?: boolean;\n    subscriber?: unknown;\n  }): Promise<StreamSubscriptionHandle>;\n}\n\n/** Window for bounded stream reads. Offsets are exclusive: `afterOffset: 10` starts at 11. */\nexport type StreamEventReadInput = {\n  /** Exclusive lower bound. Defaults to 0. */\n  afterOffset?: number;\n  /** Exclusive upper bound. Omit/null to read through the current tail. */\n  beforeOffset?: number | null;\n  /** Event types to include. Omit or include \"*\" for all; [] matches none. */\n  eventTypes?: readonly string[];\n  /** Page size, 1-500. Defaults to 500. */\n  limit?: number;\n};\n\n/**\n * Stateful page reader for one stream read window.\n *\n * This is not a live subscription; [] means \"caught up for now\". Dispose it\n * when finished (`using pager = stream.readEvents(...)`).\n */\nexport interface StreamEventPager extends Disposable {\n  /** Returns [] when no newer matching page is currently available. */\n  next(): Promise<StreamEvent[]>;\n}\n\n/** Repo catalog for either a project or the deployment-wide global scope. */\nexport interface RepoCollection extends Describable {\n  create(input: { path: string }): Promise<Repo>;\n  get(path: string): Repo;\n}\n\n/** Project-scoped repo catalog with reduced-state listing. */\nexport interface ProjectRepoCollection extends RepoCollection {\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Git-backed repo capability used by project workers and dynamic worker refs. */\nexport interface Repo extends Describable {\n  commitFiles(input: CommitRepoFilesInput): Promise<CommitRepoFilesResult>;\n  create(): Promise<Repo>;\n  /**\n   * Safely replace text in one committed file and commit the result. The\n   * `oldString` must match exactly once unless `replaceAll` is true.\n   */\n  edit(input: EditRepoFileInput): Promise<EditRepoFileResult>;\n  /** All committed file paths at HEAD. */\n  listFiles(): Promise<{ commitOid: string; paths: string[] }>;\n  processor: StreamProcessorRpc<RepoProcessorState>;\n  /** Committed file contents at HEAD; null when the path does not exist. */\n  readFile(input: { path: string }): Promise<{\n    commitOid: string;\n    content: string;\n    path: string;\n  } | null>;\n  whoami(): Promise<string>;\n}\n\n/**\n * Catalog of sandboxes within one project.\n *\n * A sandbox is addressed by its FULL path, which always lives under\n * `/sandboxes/` — the same domain-prefix convention as `/secrets/...` and\n * `/repos/...`: an agent's sandbox is the agent path under the Cloudflare\n * provider segment (`/sandboxes/cloudflare/agents/...`, exposed as\n * `itx.sandbox` in that agent's scope), and standalone sandboxes\n * conventionally live under `/sandboxes/cloudflare/<anything>`.\n * Getting a sandbox is cheap and does not start a container; the first\n * command does, and the sandbox sleeps again after idle.\n */\nexport interface SandboxCollection extends Describable {\n  get(path: string): Promise<CloudflareSandbox>;\n}\n\n/**\n * One Cloudflare Sandbox: the bare `@cloudflare/sandbox` Durable Object stub,\n * nothing wrapped on top. Whatever the installed SDK exposes is callable —\n * `exec(command)`, `readFile`/`writeFile`/`listFiles`, `startProcess`,\n * `gitCheckout`, `exposePort`, `destroy()`, … — so this contract deliberately\n * does not re-declare that surface (same stance as {@link McpClientRpc}); see\n * https://developers.cloudflare.com/sandbox/ for the API. One addition: the\n * project repo is ALWAYS checked out at `/workspace/repos/project`, which is\n * also the default working directory — a bare `exec(\"ls\")` lists the project.\n * Every command awaits that provisioning internally, so no explicit\n * `ensureProjectRepo()` is needed first (it stays available to await the\n * checkout deterministically).\n */\nexport type CloudflareSandbox = object;\n\n/** Secret catalog within one project. */\nexport interface SecretCollection extends Describable {\n  get(path: string): Secret;\n  list(): Promise<StreamListItem[]>;\n}\n\n/** Path-addressed secret capability. Secret material has no public read API. */\nexport interface Secret extends Describable {\n  describe(): Promise<SecretDescription>;\n  fetch(req: Request): Promise<Response>;\n  processor: StreamProcessorRpc<SecretDescription>;\n  update(input: SecretUpdateInput): Promise<StreamEvent>;\n}\n\nexport type SecretUpdateInput = {\n  egress?: { urls: string[] };\n  material?: string;\n};\n\nexport type SecretDescription = {\n  audit: {\n    lastUsedAt?: string;\n    lastUsedBy?: string;\n    lastUsedUrl?: string;\n    usedCount: number;\n  };\n  egress: { urls: string[] };\n  hasMaterial: boolean;\n};\n\nexport type StreamListItem = {\n  createdAt: string;\n  path: string;\n};\n\nexport type ProjectCustomDomainStatus =\n  | \"requested\"\n  | \"provisioning\"\n  | \"pending_validation\"\n  | \"active\"\n  | \"failed\"\n  | \"removing\";\n\nexport type ProjectCustomDomainValidationRecord = {\n  name: string;\n  status: string | null;\n  value: string;\n};\n\nexport type ProjectCustomDomainDnsRecord = {\n  name: string;\n  value: string;\n};\n\nexport type ProjectCustomDomain = {\n  certificateDelegationCname: ProjectCustomDomainDnsRecord | null;\n  cloudflareHostnameId: string | null;\n  createdAt: string;\n  error: string | null;\n  hostname: string;\n  hostnameStatus: string | null;\n  ownershipVerification: { name: string; value: string } | null;\n  sslStatus: string | null;\n  status: ProjectCustomDomainStatus;\n  updatedAt: string;\n  validationRecords: ProjectCustomDomainValidationRecord[];\n  wildcard: boolean;\n};\n\nexport type ProjectProcessorState = {\n  agents: StreamListItem[];\n  createRequest: { projectId: string; slug: string } | null;\n  created: boolean;\n  customDomains: ProjectCustomDomain[];\n  onboardingActive: boolean;\n  onboardingCompletedAt: string | null;\n  repos: StreamListItem[];\n  secrets: StreamListItem[];\n  streams: StreamListItem[];\n};\n\nexport type AgentProcessorState = {\n  currentRequest:\n    | { phase: \"scheduled\"; requestId: string; scheduledOffset: number }\n    | { phase: \"requested\"; llmRequestId: number }\n    | null;\n  history: Array<{ role: \"user\" | \"assistant\"; content: string }>;\n  llmConfig: { model: string };\n  llmProvider: \"cloudflare-ai\" | \"openai-ws\";\n  pendingTriggerOffset: number | null;\n  inProgressScriptExecutions: Array<{\n    code: string;\n    executionId: string;\n    requestedOffset: number;\n    startedAt: string;\n  }>;\n  scriptExecutionsCompleted: string[];\n  systemPrompt: string;\n};\n\nexport type RepoProcessorState = {\n  artifactName: string | null;\n  created: boolean;\n  defaultBranch: string | null;\n  initialized: boolean;\n  remote: string | null;\n};\n\n// NOTE: there is deliberately no public secret processor state type: a\n// secret's public live state IS its SecretDescription. The internal fold\n// carries the encrypted material; the DO's processor facade projects it away\n// (write-only material) before anything crosses the RPC boundary.\n\nexport type ProcessorSnapshot<State> = {\n  offset: number;\n  state: State;\n};\n\n/**\n * Live handle for one `onStateChange` subscription.\n *\n * `ping()` is the liveness probe: `true` while the subscription is still\n * registered on the live processor, `false` once it was dropped (delivery\n * failure, explicit unsubscribe). The call REJECTS when the hosting Durable\n * Object incarnation is gone. For a subscriber, `false` and a rejection mean\n * the same thing: re-subscribe. Pushes stop silently when a DO restarts or a\n * transport half-opens, so a periodic ping is how a client turns \"silently\n * stale\" into \"detectably dead\".\n */\nexport type ProcessorStateSubscriptionHandle = Disposable & {\n  ping(): boolean | Promise<boolean>;\n  unsubscribe(): void;\n};\n\nexport interface StreamProcessorRpc<State = unknown> {\n  getRuntimeState(): Promise<ProcessorRuntimeState<State>>;\n  /**\n   * Server-push of the processor's reduced state. The callback receives the\n   * durable checkpoint `{ offset, state }` — offset-carrying so clients can\n   * commit pushes and `snapshot()` reads monotonically against each other —\n   * once immediately on subscribe (current state IS the first paint) and then\n   * after every checkpointed batch that changed state.\n   */\n  onStateChange(\n    cb: (snapshot: ProcessorSnapshot<State>) => unknown,\n  ): Promise<ProcessorStateSubscriptionHandle>;\n  snapshot(): Promise<ProcessorSnapshot<State>>;\n  waitUntilEvent(input: { offset: number; timeoutMs?: number }): Promise<void>;\n}\n\n/**\n * Per-stub dispatch options for `DynamicWorkerCollection.get`.\n *\n * `flattenNestedPaths` mirrors `provideCapability`: dotted calls on the stub\n * become ONE `invokeCapability({ path, args })` call that the worker's own\n * `invokeCapability` method dispatches in userspace (one RPC per call),\n * instead of the default member-by-member replay on the entrypoint.\n * `buildBudgetMs` bounds how long a call waits on a cold source build; past\n * it the call fails with an error whose `name` is\n * `\"WorkerBuildInProgressError\"` — the NAME is the contract (it survives\n * Workers RPC; class identity does not), so userspace matches\n * `error.name === \"WorkerBuildInProgressError\"` to render its own building\n * page (the seeded template's router does exactly this).\n */\nexport type DynamicWorkerDispatchOptions = {\n  buildBudgetMs?: number;\n  flattenNestedPaths?: boolean;\n};\n\n/** Capability-tree entry point for ad-hoc project-scoped worker refs. */\nexport interface DynamicWorkerCollection extends Describable {\n  get<T extends object = Record<string, unknown>>(\n    ref: DynamicWorkerRef,\n    options?: DynamicWorkerDispatchOptions,\n  ): DynamicWorkerCapability<T>;\n}\n\n/** Live replacement for project egress. It sees getSecret(...) placeholders, never material. */\nexport type ProjectEgressInterceptor = (req: Request) => Promise<Response>;\n\n/** Disposable handle for one live project egress interception. */\nexport interface ProjectEgressIntercept extends Disposable {\n  release(): Promise<void>;\n}\n\n/**\n * Project-owned egress facet.\n *\n * `fetch` is the explicit outbound door. Dynamic workers' bare `fetch()` uses\n * the same project egress path through the WorkerEntrypoint gateway.\n *\n * `intercept` installs one live runtime replacement on the Project Durable\n * Object. Last writer wins; disposing or releasing the handle clears only the\n * interceptor it installed if it is still current.\n */\nexport interface ProjectEgress extends Describable {\n  fetch(req: Request): Promise<Response>;\n  intercept(handler: ProjectEgressInterceptor): Promise<ProjectEgressIntercept>;\n}\n\n/** What a project itx's `__describe()` returns: the Description convention plus identity and the capability inventory. */\nexport type ProjectDescription = Description & {\n  capabilities: CapabilityDescription[];\n  name: string;\n  projectId: string;\n};\n\nexport type CapabilityDescription = {\n  instructions?: string;\n  path: string[];\n  providedAtOffset?: number;\n  /**\n   * The itx scope path this capability is declared at (`\"/\"`, `\"/agents/bla\"`, …).\n   * Set when a scope reports capabilities it inherited from an enclosing scope,\n   * so the reader can tell a local mount from an inherited one. Absent on\n   * built-ins (they exist at every scope).\n   */\n  scope?: string;\n  type: \"builtin\" | \"live\" | \"itx-expression\";\n  types?: string;\n};\n\nexport interface OpenApiCollection extends Describable {\n  connect(input: OpenApiConnectInput): Promise<OpenApiRpc>;\n}\n\nexport type OpenApiConnectInput = {\n  baseUrl?: string;\n  headers?: Record<string, string>;\n  specUrl: string;\n};\n\nexport type OpenApiRpc = object;\n\nexport interface McpClientCollection extends Describable {\n  connect(input: McpClientConnectInput): Promise<McpClientRpc>;\n  /**\n   * The public Exa MCP server (https://mcp.exa.ai/mcp), pre-connected for every\n   * project: web search and page reading as flat tool calls.\n   * `itx.mcp.exa.web_search_exa({ query, numResults })` searches the web;\n   * `itx.mcp.exa.web_fetch_exa({ urls, maxCharacters })` reads pages as markdown.\n   */\n  exa: McpClientRpc;\n}\n\nexport type McpClientConnectInput = {\n  headers?: Record<string, string>;\n  timeoutMs?: number;\n  url: string;\n};\n\nexport type McpClientRpc = object;\n\n/**\n * Read-only catalogue of known-good itx script snippets — the project-context\n * entries of the REPL \"Examples\" panel. `list()` returns every entry without\n * its code; `get({ id })` returns one entry with the full script body. Agents\n * copy working patterns from here instead of guessing at the surface.\n * Session-level examples (whoami, projects.list) are excluded: they run\n * against the OS Session `authenticate()` returns, which an itx holder does\n * not have.\n */\nexport interface ItxExampleCatalog extends Describable {\n  get(input: { id: string }): Promise<ItxExampleWithCode>;\n  list(): Promise<ItxExampleSummary[]>;\n}\n\nexport type ItxExampleSummary = {\n  description: string;\n  id: string;\n  title: string;\n};\n\nexport type ItxExampleWithCode = ItxExampleSummary & { code: string };\n\n/**\n * The host surface for one capability scope: the durable dynamic-capability\n * table and script journal at one path inside a project (backed by one\n * CapabilityHostDurableObject). Mounting is always local to the host's own\n * scope; reads (`invokeCapability`, `__describe`) chain up through\n * enclosing scopes, so a project-root mount is visible from every scope.\n */\nexport interface CapabilityHost {\n  /** Includes `capabilities`: everything reachable at this scope — own mounts plus inherited ones, tagged with their declaring scope. */\n  __describe(): Promise<Description & { capabilities: CapabilityDescription[]; path: string }>;\n  /** The scope path this host fronts: `\"/\"` is the project root, `/agents/bla` an agent. */\n  path: string;\n  /** Explicit dynamic dispatch; the dotted-path fallback (`itx.foo.bar(...)`) compiles to exactly this call. */\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  provideCapability(input: ProvideCapabilityInput): Promise<CapabilityProvision>;\n  revokeCapability(input: RevokeCapabilityInput): Promise<void>;\n  runScript(code: string): Promise<{\n    completedEvent: StreamEvent;\n    executionId: string;\n    result: unknown;\n  }>;\n}\n\n/** Capability hosts by scope path within one project (`itx.capabilityHosts`). */\nexport interface CapabilityHostCollection extends Describable {\n  get(path: string): CapabilityHost;\n}\n\n/**\n * Ownership handle for one mounted capability.\n *\n * The offset identifies the exact `capability-provided` event, which lets\n * disposal or explicit revoke remove this mount without racing a newer mount at\n * the same path.\n */\nexport interface CapabilityProvision extends Describable, Disposable {\n  readonly path: string[];\n  readonly providedAtOffset: number;\n  revoke(): Promise<void>;\n}\n\n/** Append input before the stream assigns offset and timestamp. */\nexport type StreamEventSource = {\n  processor?: {\n    slug: string;\n    version: string;\n  };\n  crossPost?: {\n    ruleId: string;\n    from: {\n      createdAt: string;\n      offset: number;\n      path: string;\n      projectId: string | null;\n      type: string;\n    };\n  };\n};\n\nexport type StreamEventInput = {\n  type: string;\n  payload?: Record<string, unknown>;\n  metadata?: Record<string, unknown>;\n  source?: StreamEventSource;\n  idempotencyKey?: string;\n};\n\n/** Durable stream event after commit. */\nexport type StreamEvent = StreamEventInput & {\n  createdAt: string;\n  offset: number;\n};\n\n/** Stable identity for one stream subscription connection. */\nexport type SubscriptionKey = string;\n\n/**\n * Batch delivered to stream processors and live subscribers.\n *\n * Kept named because callback retention, processor hosts, and tests all depend\n * on the same cross-RPC batch envelope.\n */\nexport type StreamEventBatch = {\n  projectId: string | null;\n  path: string;\n  events: StreamEvent[];\n  streamMaxOffset: number;\n  state: unknown;\n};\n\n/**\n * Callback invoked by the stream pump for each delivered batch.\n *\n * It stays as a named type because Workers RPC callback lifecycle helpers need\n * to duplicate, retain, and dispose exactly this callback shape.\n */\nexport type ProcessEventBatch = (batch: StreamEventBatch) => unknown;\n\n/** Serializable snapshot plus optional live runtime debug state for a processor. */\nexport type ProcessorRuntimeState<State = unknown> = {\n  snapshot: { offset: number; state: State };\n  runtime?: Record<string, unknown>;\n};\n\n/**\n * Optional runtime-state callback exposed by a hosted processor.\n *\n * It accepts sync or async implementations because local processors can return\n * immediately, while RPC-backed processors may need an async round trip.\n */\nexport type GetProcessorRuntimeState = () => ProcessorRuntimeState | Promise<ProcessorRuntimeState>;\n\n/**\n * Live subscription handle returned by `Stream.subscribe`.\n *\n * `ping()` mirrors {@link ProcessorStateSubscriptionHandle.ping}: `true` while\n * the connection is still open on the live stream, `false` after it closed\n * (replaced, delivery failure, unsubscribe); it rejects when the stream's\n * Durable Object incarnation is gone. Either non-`true` outcome means the\n * subscriber should re-subscribe.\n */\nexport type StreamSubscriptionHandle = Disposable & {\n  subscriptionKey: SubscriptionKey;\n  streamMaxOffset: number;\n  ping(): boolean | Promise<boolean>;\n  unsubscribe(): void;\n};\n\n/**\n * One repo file mutation.\n *\n * Kept named because public `Repo.commitFiles`, input parsing, and artifact\n * commit implementation all validate the same command shape.\n */\nexport type RepoFileChange =\n  | {\n      path: string;\n      content: string;\n    }\n  | {\n      path: string;\n      delete: true;\n    };\n\n/** Command object for committing a batch of repo file mutations. */\nexport type CommitRepoFilesInput = {\n  author?: { email: string; name: string };\n  branch?: string;\n  changes: RepoFileChange[];\n  message: string;\n};\n\n/** Result returned after a repo commit attempt, including no-op commits. */\nexport type CommitRepoFilesResult = {\n  branch: string;\n  changedPaths: string[];\n  commitOid: string;\n  noChanges: boolean;\n};\n\n/** Command object for a coding-agent-style exact string edit. */\nexport type EditRepoFileInput = {\n  author?: { email: string; name: string };\n  branch?: string;\n  message: string;\n  newString: string;\n  oldString: string;\n  path: string;\n  replaceAll?: boolean;\n};\n\n/** Result returned after an exact string edit commit attempt. */\nexport type EditRepoFileResult = CommitRepoFilesResult & {\n  occurrenceCount: number;\n  path: string;\n};\n\n/** Dynamic invocation envelope used by flattened live capabilities. */\nexport type FlattenedCapabilityInvocation = {\n  args: unknown[];\n  flattenNestedPath?: boolean;\n  path: string[];\n};\n\n/** Target shape for a live capability that wants to receive flattened paths. */\nexport type FlattenedCapabilityTarget = {\n  invokeCapability(input: FlattenedCapabilityInvocation): unknown;\n};\n\n/** Durable expression over the project ITX surface. */\nexport type ItxExpressionStep = string | [method: string, ...args: unknown[]];\nexport type ItxExpression = ItxExpressionStep[];\n\n/** Capability recipe accepted by `provideCapability`. */\nexport type ProvideCapabilityInput =\n  | {\n      capability: unknown;\n      flattenNestedPaths?: false;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      capability: FlattenedCapabilityTarget;\n      flattenNestedPaths: true;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      expression: ItxExpression;\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"itx-expression\";\n      types?: string;\n    };\n\n/** Event payload stored when a capability is mounted on an ITX stream. */\nexport type CapabilityProvidedPayload =\n  | {\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"live\";\n      types?: string;\n    }\n  | {\n      expression: ItxExpression;\n      flattenNestedPaths?: boolean;\n      instructions?: string;\n      path: string[];\n      type: \"itx-expression\";\n      types?: string;\n    };\n\n/** Reduced capability table row: payload plus the providing event offset. */\nexport type CapabilityRecord = CapabilityProvidedPayload & {\n  providedAtOffset: number;\n};\n\n/** Revoke command for the current mount at a path or one exact mount offset. */\nexport type RevokeCapabilityInput = {\n  path: string[];\n  providedAtOffset?: number;\n};\n\n/**\n * Credentials accepted by `UnauthenticatedOs.authenticate`.\n *\n * - `from-server-cookie` — the browser lane: the deployment's admin cookie or\n *   the signed-in user's session cookie riding the WebSocket handshake.\n * - `bearer` — an auth access token presented as RPC data.\n * - `admin-secret` — the deployment admin API secret (CLI / tooling / e2e).\n * - `impersonate` — admin-secret-gated fake principal, for test suites that\n *   exercise per-project confinement without minting real users.\n */\nexport type ItxAuthCredentials =\n  | { type: \"from-server-cookie\" }\n  | { type: \"bearer\"; token: string }\n  | { type: \"admin-secret\"; secret: string }\n  | { type: \"impersonate\"; secret: string; token: ItxAuthToken };\n\n/** Principal shape for `impersonate` credentials. */\nexport type ItxAuthToken =\n  | { type: \"admin\"; principal?: string }\n  | { type: \"user\"; principal: string; projectScopes: string[] };\n\n/** Authority object carried by server-side RPC target instances. */\nexport interface ItxAuth {\n  readonly principal: string;\n  isAdmin(): boolean;\n  canAccessProject(projectId: string): boolean;\n  assertCanAccessProject(projectId: string | null): void;\n  listAccessibleProjects(): string[];\n  /**\n   * Async access check that may consult the project directory (source of\n   * truth) when synchronous claims miss — see the auth adapter. Optional so\n   * in-process trusted contexts stay trivially constructible.\n   */\n  ensureCanAccessProject?(projectId: string): Promise<void>;\n}\n\n/**\n * Where a dynamic worker's source files come from.\n *\n * `inline` supplies the file map directly — the primitive behind run-script and\n * worker-backed provided capabilities where the caller hands over a small\n * TypeScript entry file, helpers, and optionally a `package.json`. `repo` names\n * a project repo snapshot: a branch (late-bound, so future commits affect the\n * next use) or a pinned commit, narrowed by include/exclude glob masks so a\n * large repo does not become build input by default.\n */\nexport type WorkerFileSource =\n  | {\n      type: \"inline\";\n      files: Record<string, string>;\n    }\n  | {\n      type: \"repo\";\n      repoPath: string;\n      /**\n       * Defaults to the repo's default branch when omitted. A pinned commit\n       * may name the branch it lives on — clones are single-branch, so an\n       * off-default-branch commit is unreachable without it.\n       */\n      ref?: { branch: string } | { commitOid: string; branch?: string };\n      include?: string[];\n      exclude?: string[];\n    };\n\n/** Loader names accepted by Cloudflare's worker bundler `loader` option. */\nexport type WorkerBundlerLoader =\n  | \"js\"\n  | \"jsx\"\n  | \"ts\"\n  | \"tsx\"\n  | \"json\"\n  | \"css\"\n  | \"text\"\n  | \"binary\"\n  | \"base64\"\n  | \"dataurl\";\n\n/**\n * Build options for a dynamic worker.\n *\n * This mirrors Cloudflare's `CreateWorkerOptions` from\n * `@cloudflare/worker-bundler` minus `files` (OS supplies files from the\n * selected {@link WorkerFileSource}) — deliberately not a parallel option\n * language (drift fails typecheck via the assignability pin in\n * domains/workers/schemas.ts; this file stays import-free because it is\n * published verbatim as the itx types surface). `bundle: false` is allowed; the invariant is one OS\n * materialization pipeline, not one bundled output file. When the file map has\n * a `package.json` with dependencies, the bundler installs them from the npm\n * registry at build time.\n */\nexport type WorkerBuildOptions = {\n  /** Entry point file path relative to the source root (e.g. \"worker.ts\"). */\n  entryPoint?: string;\n  /** Bundle all dependencies into a single output file. Default: true. */\n  bundle?: boolean;\n  /** Modules kept external (\"cloudflare:*\" always is). */\n  externals?: string[];\n  /** Target environment. Default: \"es2022\". */\n  target?: string;\n  minify?: boolean;\n  sourcemap?: boolean;\n  /** npm registry URL for dependency installs. */\n  registry?: string;\n  jsx?: \"transform\" | \"preserve\" | \"automatic\";\n  jsxImportSource?: string;\n  define?: Record<string, string>;\n  loader?: Record<string, WorkerBundlerLoader>;\n  conditions?: string[];\n  virtualModules?: Record<string, string>;\n};\n\n/**\n * Declarative source for a dynamic worker: an orthogonal file source plus\n * Cloudflare-compatible build options.\n *\n * Materialization resolves `files` to a file map and builds it through\n * Cloudflare's worker bundler; the loader-ready output is cached by a\n * deterministic build key, so the same source+options never builds twice.\n */\nexport type DynamicWorkerSource = {\n  files: WorkerFileSource;\n  options?: WorkerBuildOptions;\n};\n\ntype DynamicWorkerRefBase = {\n  /**\n   * ITX scope path for the worker's `env.ITX` binding and for stateful worker\n   * Durable Object names. This is intentionally not the mounted capability path:\n   * one worker can be mounted at `db`, `counter`, etc. while all events still\n   * belong to the host stream path.\n   */\n  path: string;\n  source: DynamicWorkerSource;\n};\n\n/**\n * Stateless workers are WorkerEntrypoint exports loaded directly from source.\n *\n * `props` are passed to `worker.getEntrypoint(name, { props })` and appear as\n * `this.ctx.props` inside the exported WorkerEntrypoint. They deliberately live\n * only on stateless refs: Durable Object facets are started with\n * `ctx.facets.get(name, () => ({ class, id? }))`, which does not accept\n * WorkerEntrypoint-style props.\n */\nexport type StatelessDynamicWorkerRef = DynamicWorkerRefBase & {\n  type: \"stateless\";\n  entrypoint?: string;\n  props?: Record<string, JsonValue>;\n};\n\n/**\n * Stateful workers are Durable Object class exports hosted by\n * `StatefulWorkerDurableObject`.\n *\n * `durableWorkerKey` is the durable identity under `{ projectId, path }`. It is\n * not a source cache key: source changes deliberately affect the next use of the\n * same durable worker identity.\n */\nexport type StatefulDynamicWorkerRef = DynamicWorkerRefBase & {\n  type: \"stateful\";\n  className: string;\n  durableWorkerKey: string;\n  /**\n   * What a call does when the worker's source changed since the running\n   * version. `\"block\"` (default) waits for the rebuild — commit-then-call\n   * sees the new code. `\"stale-while-rebuild\"` keeps answering with the\n   * running version and swaps to the new build in the background: better\n   * availability, but the next few calls after a commit may see old code.\n   * The policy rides the REF, not the durable identity — callers sharing one\n   * `durableWorkerKey` should agree on it (and on `source`), or each call\n   * flips the facet to its own version.\n   */\n  updatePolicy?: \"block\" | \"stale-while-rebuild\";\n};\n\n/** Worker recipe accepted by `workers.get` and worker-backed capabilities. */\nexport type DynamicWorkerRef = StatelessDynamicWorkerRef | StatefulDynamicWorkerRef;\n\n/** Dynamic worker RPC stub plus the disposal operation owned by the caller. */\nexport type DynamicWorkerCapability<T extends object = Record<string, unknown>> = T & Disposable;\n\n/**\n * Slack Web API surface exposed by the seeded project worker\n * (`itx.worker.slack.chat.postMessage({...})`).\n *\n * The seeded repo implements this in userland with the real `@slack/web-api`\n * package (installed by the worker build pipeline from its `package.json`), so\n * any nested Web API method family resolves — the index signature reflects\n * that this tree is as wide as the SDK's.\n */\nexport interface ProjectWorkerSlack {\n  chat: {\n    postMessage(input: Record<string, unknown>): Promise<Record<string, unknown>>;\n  } & Record<string, unknown>;\n  [family: string]: unknown;\n}\n\n/**\n * Default seeded project worker contract.\n *\n * This documents the reference repo's `worker.ts` only. Arbitrary dynamic\n * workers should be typed by callers through `workers.get<T>(ref)`. The\n * platform dispatches to it with flattened paths, so the worker implements\n * `invokeCapability` in userspace and every dotted call — including any\n * nested `slack.*` Web API family — is one RPC.\n */\nexport interface ProjectWorker {\n  fetch(req: Request): Promise<Response>;\n  invokeCapability(input: { args?: unknown[]; path: string[] }): Promise<unknown>;\n  processEvent(input: { event: StreamEvent }): Promise<void>;\n  slack: ProjectWorkerSlack;\n}\n\n/** JSON subset accepted by WorkerEntrypoint props and script results. */\nexport type JsonValue =\n  | string\n  | number\n  | boolean\n  | null\n  | JsonValue[]\n  | { [key: string]: JsonValue };\n\n/**\n * Minimal ExecutionContext shape the RPC adapter layer needs.\n *\n * This stays at the bottom because it is server-side plumbing, not part of the\n * client-facing ITX contract. It is exported only so domain hosts can inject\n * project/agent capability targets without importing the full worker module.\n */\nexport type CfExecutionContext = {\n  exports: ExecutionContext[\"exports\"];\n  /** Required, not optional: budget-expired dynamic worker builds are handed\n   * to it (worker-loader.ts withBuildBudget) — a silent no-op here would\n   * strand every cold build the caller gave up on. Hosts without a real\n   * runtime hook must still supply an explicit function. */\n  waitUntil: ExecutionContext[\"waitUntil\"];\n};\n// \u0063odegen:end\n" },
   { path: "slack.config.ts", content: "/**\n * Configuration for the worker's Slack Web API surface (see `slack` in\n * worker.ts). Commit changes to this file to point the SDK at your workspace:\n * set `token` to a bot token, and leave `slackApiUrl` null for the real Slack\n * API (override it only for mocks or staging proxies).\n */\nexport const slackConfig: {\n  slackApiUrl: string | null;\n  token: string | null;\n} = {\n  slackApiUrl: null,\n  token: null,\n};\n" },
   { path: "tsconfig.json", content: "{\n  \"include\": [\"**/*.ts\"],\n  \"compilerOptions\": {\n    \"target\": \"ES2024\",\n    \"module\": \"ESNext\",\n    \"moduleResolution\": \"bundler\",\n    \"allowImportingTsExtensions\": true,\n    \"noEmit\": true,\n    \"lib\": [\"ES2024\"],\n    \"types\": [\"@cloudflare/workers-types\"],\n    \"skipLibCheck\": true,\n    \"strict\": true\n  }\n}\n" },
   { path: "worker.ts", content: "import { WorkerEntrypoint } from \"cloudflare:workers\";\nimport { WebClient } from \"@slack/web-api\";\nimport type { DynamicWorkerRef, ProjectRpcTarget, StreamEvent } from \"./sdk.ts\";\nimport { slackConfig } from \"./slack.config.ts\";\n\n/** Bindings the platform supplies to every project worker. */\ntype ProjectWorkerEnv = {\n  ITX: { get(): Promise<ProjectRpcTarget> };\n};\n\n// The root project worker is a small ROUTER over the project's apps. Each app\n// is its own repo-backed dynamic worker built from this repo (multi-file\n// TypeScript — the build pipeline bundles the masked file snapshot); ingress\n// selects one via the trusted x-iterate-app header (hosts like\n// hello--<slug>.<base> or <app>.<custom-hostname>). Requests with no app\n// selected get the static homepage below.\nconst APPS = {\n  hello: {\n    type: \"stateless\",\n    path: \"/\",\n    source: {\n      files: { type: \"repo\", repoPath: \"/\", include: [\"apps/hello/**\", \"sdk.ts\"] },\n      options: { entryPoint: \"apps/hello/worker.ts\" },\n    },\n  },\n  counter: {\n    type: \"stateful\",\n    path: \"/\",\n    className: \"CounterApp\",\n    durableWorkerKey: \"app-counter\",\n    source: {\n      files: { type: \"repo\", repoPath: \"/\", include: [\"apps/counter/**\"] },\n      options: { entryPoint: \"apps/counter/worker.ts\" },\n    },\n  },\n} satisfies Record<string, DynamicWorkerRef>;\n\nexport default class ProjectWorker extends WorkerEntrypoint<ProjectWorkerEnv> {\n  async fetch(req: Request): Promise<Response> {\n    const appSlug = req.headers.get(\"x-iterate-app\");\n    if (appSlug) {\n      const ref = Object.hasOwn(APPS, appSlug) ? APPS[appSlug as keyof typeof APPS] : undefined;\n      if (!ref) return new Response(`unknown app: ${appSlug}`, { status: 404 });\n      const project = await this.env.ITX.get();\n      try {\n        // Workers RPC: await the capability before calling through it.\n        const app = await project.workers.get<{ fetch(req: Request): Promise<Response> }>(ref, {\n          buildBudgetMs: 15_000,\n        });\n        return await app.fetch(req);\n      } catch (error) {\n        // Each app is its own build (distinct include mask), so its first use\n        // after a commit can be a cold build. Past the budget the platform\n        // throws this named error (name survives Workers RPC) while the build\n        // finishes into the cache; refreshes then hit it.\n        if ((error as { name?: string } | null)?.name !== \"WorkerBuildInProgressError\") throw error;\n        return new Response(\n          `<!doctype html><html><body><main><p>Building ${appSlug}… this page refreshes automatically.</p></main></body></html>`,\n          { headers: { \"content-type\": \"text/html; charset=utf-8\", refresh: \"3\" }, status: 503 },\n        );\n      }\n    }\n\n    // The seeded homepage is a static page linking to the apps. Apps live\n    // on their own hosts: the current host prefixed with \"<app>--\" (e.g.\n    // counter--<slug>.<base>), so the links derive from the request URL.\n    const url = new URL(req.url);\n    const appLinks = Object.entries(APPS)\n      .map(([slug, ref]) => {\n        const href = `${url.protocol}//${slug}--${url.host}/`;\n        return `<li><a href=\"${href}\">${slug}</a> (${ref.type})</li>`;\n      })\n      .join(\"\\n\");\n    return new Response(\n      `<!doctype html>\n        <html>\n          <body>\n            <main>\n              <p>Hello from your Iterate project worker.</p>\n              <ul>${appLinks}</ul>\n              <p>Edit worker.ts in the project repo to change this.</p>\n            </main>\n          </body>\n        </html>`,\n      { headers: { \"content-type\": \"text/html; charset=utf-8\" } },\n    );\n  }\n\n  processEvent(input: { event: StreamEvent }): void {\n    console.log(\"project worker processed\", input.event.type);\n  }\n\n  /**\n   * The platform dispatches dotted calls on this worker as ONE flattened\n   * `invokeCapability({ path, args })` call, and this userspace method walks\n   * the path over the worker itself. That is what lets the `slack` getter\n   * below hand back the raw SDK client: nothing ever crosses RPC except the\n   * final method's arguments and result, so\n   * `itx.worker.slack.chat.postMessage({...})` — or any nested Web API\n   * family — is a single round trip into plain userland code.\n   */\n  async invokeCapability({ args = [], path }: { args?: unknown[]; path: string[] }) {\n    let receiver: unknown = this;\n    for (const segment of path.slice(0, -1)) {\n      receiver = await Reflect.get(Object(receiver), segment);\n    }\n    const method = path.at(-1)!;\n    const handler = Reflect.get(Object(receiver), method);\n    if (typeof handler !== \"function\") {\n      throw new Error(`\"${path.join(\".\")}\" is not a method on this project worker`);\n    }\n    return await Reflect.apply(handler, receiver, args);\n  }\n\n  /**\n   * Slack Web API surface: the real `@slack/web-api` SDK from package.json,\n   * configured by committing slack.config.ts. Only ever reached through the\n   * userspace `invokeCapability` walk above, so the client needs no RPC-safe\n   * projection.\n   */\n  get slack(): WebClient {\n    const client = new WebClient(slackConfig.token ?? undefined, {\n      ...(slackConfig.slackApiUrl === null ? {} : { slackApiUrl: slackConfig.slackApiUrl }),\n    });\n    // The SDK's axios defaults to its node-http adapter, whose response\n    // handling hangs under the Workers runtime; the fetch adapter rides the\n    // platform's native fetch (and therefore project egress) instead.\n    (client as unknown as { axios: { defaults: { adapter: string } } }).axios.defaults.adapter =\n      \"fetch\";\n    return client;\n  }\n}\n" },
